% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_spatialRaster.R
\name{add_proj4_and_epsg_to_spatialReferenceDictionary}
\alias{add_proj4_and_epsg_to_spatialReferenceDictionary}
\title{Add proj4 and EPSG codes to the EML spatial reference dictionary}
\usage{
add_proj4_and_epsg_to_spatialReferenceDictionary()
}
\value{
(xml_document, xml_node) The EML spatial reference dictionary proj4 and
    EPSIG codes added when matches are available using the tags:
    \item{proj4}{The corresponding proj4 code}
    \item{epsg}{The corresponding EPSG code}
}
\description{
The EML spatial reference dictionary contains coordinate reference system 
(CRS) names that are loosely based on ESRI but are not common among other 
spatial software libraries. This function adds the exact ESRI names and 
corresponding proj4 and EPSG codes to improve metadata links.
}
\details{
This function uses spatialreference.org to translate between CRS names
    and codes.
    
    This function assumes ...
}
\note{
This function may not match all CRS names in the EML dictionary.
}
