% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate_eml.R
\name{annotate_eml}
\alias{annotate_eml}
\title{Annotate an EML document}
\usage{
annotate_eml(
      annotations = NULL,
      eml.in = NULL,
      eml.out = NULL,
      x = NULL
    )
}
\arguments{
\item{annotations}{(character; data frame) Path to the annotations.txt metadata template 
for the EML document being annotated. Create annotations.txt with 
\code{template_annotations()}. Alternatively input to this argument
can be a data frame of the annotations.txt template.}

\item{eml.in}{(character; emld list) Path to the EML document being annotated. 
Alternatively this input can be supplied as an emld list object created 
by \code{make_eml()} or \code{EMLassemblyline::read_eml()}.}

\item{eml.out}{(character) Path to which the annotated EML should be written.}
}
\value{
If the input to \code{eml.in} is a path to an EML document (.xml) then 
    the output will be the annotated EML document written to \code{eml.out}.
    If the input to \code{eml.in} is an emld list object then the output 
    will be the annotated version of the emld list object.
}
\description{
Annotate an EML document.
}
\details{
All annotatable elements are assigned IDs and their annotations are 
    located both immediately under the parent element (subject) within
    /eml/dataset when supported and within /eml/dataset/annotations 
    node through references. This redundant approach supports variation 
    in where EML metadata consumers harvest this information and supports 
    the extensibility of annotating elements requiring 
    /eml/dataset/annotations references.
}
