% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/template_categorical_variables.R
\name{attribute_template_to_table}
\alias{attribute_template_to_table}
\title{Convert attributes file name to the corresponding data table file name}
\usage{
attribute_template_to_table(attributes.template, data.path)
}
\arguments{
\item{attributes.template}{(character) Table attributes template file name, including file extension}

\item{data.path}{(character) Path to the data directory}
}
\value{
(character) Data table file name
}
\description{
Convert attributes file name to the corresponding data table file name
}
