% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_spatialRaster.R
\name{build_raster_element}
\alias{build_raster_element}
\title{Build the <spatialRaster> elements}
\usage{
build_raster_element(r, rv = raster_var, path = path, data.path = data.path)
}
\arguments{
\item{r}{(list) A single row from the raster_template}

\item{rv}{(data.frame) The raster_var object in its entirety}

\item{path}{(character) Path to the metadata template directory.}

\item{data.path}{(character) Path to the data directory.}
}
\value{
\item{spatialRaster}{A single \code{<spatialRaster>} element (list).}
}
\description{
Build the <spatialRaster> elements
}
