% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_spatialVector.R
\name{build_shape_element}
\alias{build_shape_element}
\title{Build the <spatialVector> elements for shapefiles}
\usage{
build_shape_element(s, path = path, data.path = data.path)
}
\arguments{
\item{s}{(list) A single row from the shape_template object.}

\item{path}{(character) Path to the metadata template directory.}

\item{data.path}{(character) Path to the data directory.}
}
\value{
\item{spatialVector}{A single \code{<spatialVector>} element (list).}
}
\description{
Build the <spatialVector> elements for shapefiles
}
