% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_spatialRaster.R
\name{build_spatialRaster}
\alias{build_spatialRaster}
\title{Build spatialRaster node}
\usage{
build_spatialRaster(
  rasterFname,
  rasterDesc,
  rasterGeoDesc,
  rasterAttributes,
  rasterFactors = NULL,
  baseURL = ""
)
}
\arguments{
\item{rasterFname}{(character) Full file path}

\item{rasterDesc}{(character) Free-text descr. of raster}

\item{rasterGeoDesc}{(character) Raster geographic coverage}

\item{rasterAttributes}{(data.frame) Attributes table}

\item{rasterFactors}{(data.frame) Factor codes table}

\item{baseURL}{(character) Public URL of raster file}
}
\value{
(emld; list) EML spatialRaster node
}
\description{
Build spatialRaster node
}
\examples{
raster.attribs <- data.frame(
  attributeName = "cover_value",
  attributeDefinition = 'Shrub cover fraction (0-1)',
  unit = 'dimensionless',
  numberType = 'real',
  missingValueCode = 0,
  missingValueCodeExplanation = 'No data')

# (NOTE: These metadata don't pertain to this .tif)
covercode <- c('0' = "no data", '1' = "non-shrub", '2' = "shrub")
raster.factor <- data.frame(
  attributeName = "cover_code",
  code = names(covercode),
  definition = unname(covercode))

build_spatialRaster(
  rasterFname = "C:/Users/Colin/Documents/EDI/data_sets/fix_15/data_objects/shrub_cover.tif", 
  rasterDesc = "1-hectare JER-CDRRC shrub cover geotiff (cover fraction from 0 to 1)", 
  rasterGeoDesc = "JER and CDRRC, southern New Mexico", 
  rasterAttributes = raster.attribs, 
  rasterFactors = raster.factor,
  baseURL = "https://jornada-ldc2.jrn.nmsu.edu/tif/")

}
