% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_spatialVector.R
\name{build_vector_element}
\alias{build_vector_element}
\title{Build the <spatialVector> elements for KML files}
\usage{
build_vector_element(v, path = path, data.path = data.path)
}
\arguments{
\item{path}{(character) Path to the metadata template directory.}

\item{data.path}{(character) Path to the data directory.}

\item{s}{(list) A single row from the vector_template object.}
}
\value{
\item{spatialVector}{A single \code{<spatialVector>} element (list).}
\item{KML file}{If the original KML file has multiple layers, returns a KML file for each layer specified in vector attributes template.}
}
\description{
Build the <spatialVector> elements for KML files
}
