% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/template_categorical_variables.R
\name{catvars_from_netcdf}
\alias{catvars_from_netcdf}
\title{Create a categorical variables template for a MIME Type application/x-netcdf
data object}
\usage{
catvars_from_netcdf(data_path, attrs_tmplt, empty = FALSE)
}
\arguments{
\item{data_path}{(character) Path to the data object.}

\item{attrs_tmplt}{(data.frame) The attributes template corresponding to
the data object.}

\item{empty}{(logical) Whether to create an empty template.}
}
\value{
(data.frame or NULL) Returns a \code{data.frame} if categorical
variables are listed in \code{attrs_tmplt} corresponding categorical codes
are found in the data object, otherwise \code{NULL}.
}
\description{
Create a categorical variables template for a MIME Type application/x-netcdf
data object
}
\keyword{internal}
