% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/template_categorical_variables.R
\name{catvars_from_textcsv}
\alias{catvars_from_textcsv}
\title{Create a categorical variables template for a MIME Type text/csv data object}
\usage{
catvars_from_textcsv(data_object, attrs_tmplt, empty = FALSE)
}
\arguments{
\item{data_object}{(data.frame) The data object.}

\item{attrs_tmplt}{(data.frame) The attributes template corresponding to
\code{data_object}.}
}
\value{
(data.frame or NULL) Returns a \code{data.frame} if categorical
variables are listed in \code{attrs_tmplt} corresponding categorical codes
are found in \code{data_object}, otherwise \code{NULL}.
}
\description{
Create a categorical variables template for a MIME Type text/csv data object
}
\keyword{internal}
