% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_templates.R
\name{check_uri}
\alias{check_uri}
\title{Check resolvability of URI}
\usage{
check_uri(uri)
}
\arguments{
\item{uri}{URI to check}
}
\value{
(logical) TRUE if can be resolved, FALSE otherwise
}
\description{
Check resolvability of URI
}
\keyword{internal}
