% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eml2eal.R
\name{cnvmt}
\alias{cnvmt}
\title{Convert empty text to empty EAL text}
\usage{
cnvmt(txt)
}
\arguments{
\item{txt}{(character) Text}
}
\value{
(character) EAL representation of empty values in tabular templates
}
\description{
Convert empty text to empty EAL text
}
