% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_templates.R
\name{compile_geographic_coverage}
\alias{compile_geographic_coverage}
\title{Compile geographic coverage from multiple sources}
\usage{
compile_geographic_coverage(x)
}
\arguments{
\item{x}{(list) The data and metadata object returned by 
\code{template_arguments()}.}
}
\value{
\item{x}{(list) With geographic coverage compiled from multiple input 
    sources into the geographic_coverage template.}
}
\description{
Compile geographic coverage from multiple sources
}
\details{
Combine multiple sources of geographic coverage and remove duplicate 
    entries. This info can be supplied in the \code{geographic.coverage} and 
    \code{geographic.description} arguments of \code{make_eml()} as well as 
    in the geographic_coverage and bounding_boxes templates.
}
\keyword{internal}
