% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_templates.R
\name{compile_provenance}
\alias{compile_provenance}
\title{Compile provenance from multiple sources}
\usage{
compile_provenance(x)
}
\arguments{
\item{x}{(list) The data and metadata object returned by 
\code{template_arguments()}.}
}
\value{
\item{x}{(list) With geographic coverage compiled from multiple input 
    sources into the geographic_coverage template.}
}
\description{
Compile provenance from multiple sources
}
\details{
Combine multiple sources of provenance and remove duplicate 
    entries. This info can be supplied in the \code{provenance} 
    argument of \code{make_eml()} as well as in the provenance template.
}
