% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testthat_helpers.R
\name{copy_test_package}
\alias{copy_test_package}
\title{Copy test package files to directory path}
\usage{
copy_test_package(path)
}
\arguments{
\item{path}{(character) Directory path where test files will be copied.}
}
\value{
(character) Full paths of files copied to \code{path}.
}
\description{
Copy test package files to directory path
}
\examples{
\dontrun{
testdir <- paste0(tempdir(), "/pkg")
pkg_files <- copy_test_package(testdir)
pkg_files
unlink(testdir, recursive = TRUE, force = TRUE)
}

}
\keyword{internal}
