% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_eml.R
\name{create_attributeList}
\alias{create_attributeList}
\title{Create attributeList EML element}
\usage{
create_attributeList(x, file, file_type)
}
\arguments{
\item{x}{(named list) List of templates created with \code{template_arguments()}.}

\item{file}{(character) Name of file containing the attributes.}

\item{file_type}{(character) Type of file according to EML file field. Can be
'dataTable', 'spatialRaster', 'spatialVector' or 'otherEntity'.}
}
\value{
an eml attributeList object
}
\description{
Create attributeList EML element
}
\keyword{internal}
