% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_directory.R
\name{create_directory}
\alias{create_directory}
\title{Create directory for `EMLassemblyline` inputs and outputs}
\usage{
create_directories(path, name)
}
\arguments{
\item{path}{(character) Path to where the directory structure will be created.}

\item{dir.name}{(character) Name of parent directory. Subdirectories have predefined 
names.}
}
\value{
A directory with the following structure:
    \itemize{
        \item{`name` Parent directory with given name}
        \itemize{
            \item{`data` Where data and other entities are stored}
            \item{`scripts` Where data processing and packaging scripts are
            kept.}
            \item{`metadata_templates` Where `EMLassemblyline` template 
            files are imported and edited}
            \item{`eml` Where EML files are exported to by the 
            `EMLassemblyline`}
        }
    }
}
\description{
Create a commonly used directory structure for `EMLassemblyline` inputs 
    and outputs. You're welcome to organize content in any way you choose,
    this is just one approach.
}
\note{
Existing `dir.name`s at `path` will not be overwritten.
}
