% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_eml.R
\name{create_eml}
\alias{create_eml}
\title{Create EML and write to file}
\usage{
create_eml(path)
}
\arguments{
\item{path}{A path to the working directory containing files required by the EMLtools
package. Run the function run_guide() for instructions on how to populate this working
directory with the requisite files.}
}
\value{
A .xml file in the specified working directory. If the EML fails validation
then the file is not written.
}
\description{
A function for writting metadata in the Ecological Metadata Language
standard.
}
\seealso{
\code{\link{run_guide}} for guidance on setting up the working directory

\code{\link{write_attributes}} for writing data attributes

\code{\link{write_factors}} for writing data factors

\code{\link{copy_templates}} for creating template files
}
