% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_spatialRaster.R
\name{create_spatialRaster}
\alias{create_spatialRaster}
\title{Create EML \code{<spatialRaster>} elements}
\usage{
create_spatialRaster(
  path,
  data.path = path,
  raster_attributes = NULL,
  raster_catvars = NULL
)
}
\arguments{
\item{path}{(character) Path to the metadata template directory.}

\item{data.path}{(character) Path to the data directory.}

\item{raster_attributes}{(data.frame; optional) Alternative to the raster attributes template; enter attributes into a data.frame object and provide as an argument}

\item{raster_catvars}{(data.frame; optional) Alternative to the raster variables template; enter codes and definitions into a data.frame object and provide as an argument}
}
\value{
\item{\code{<spatialRaster>} element (list) that can be inserted into an EML file or object via \code{make_eml()}.}
}
\description{
Create EML \code{<spatialRaster>} elements
}
\examples{
\dontrun{

# Set working directory
setwd("/Users/me/Documents/data_packages/pkg_260")

# For use with raster_attributes.txt (and optionally raster_variables.txt)
spatialRaster <- create_spatialRaster(
  path = "./metadata_templates",
  data.path = "./data_objects")


# For use without a template

my_attributes <- data.frame(
  filename = 'my_raster_file.tiff',
  geoDescription = 'Greater Rio Grande valley area',
  definition = 'Normalized Difference Vegetation Index',
  unit = 'dimensionless',
  numberType = 'real')

spatialRaster <- create_spatialRaster(
  path = "./metadata_templates",
  data.path = "./data_objects",
  raster_attributes = my_attributes)
}
}
