% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_spatialRaster.R
\name{create_spatialReferenceDictionary}
\alias{create_spatialReferenceDictionary}
\title{Recreate the EML spatialReferenceDictionary and add EPSG and proj4 codes}
\usage{
create_spatialReferenceDictionary()
}
\value{
(xml_document, xml_node) The EML spatialReferenceDictionary with EPSG
    and proj4 codes added
}
\description{
Recreate the EML spatialReferenceDictionary and add EPSG and proj4 codes
}
\details{
Because the horizCoordSysDef and geogCoordSys values have an arbitrary
    format, the other components of the CRS are used as a key and means 
    of validation.
}
