% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_spatialVector.R
\name{create_spatialVector}
\alias{create_spatialVector}
\title{Create EML \code{<spatialVector>} elements}
\usage{
create_spatialVector(path, data.path = path, vector_attributes = NULL)
}
\arguments{
\item{path}{(character) Path to the metadata template directory.}

\item{data.path}{(character) Path to the data directory.}

\item{vector_attributes}{(data.frame; optional) Alternative to a vector attributes template; enter attributes into a data.frame object and provide as an argument.}
}
\value{
\item{spatialVector}{\code{<spatialVector>} element (list) that can be inserted into an EML file or object via \code{make_eml()}.}
}
\description{
Creates a list of <spatialVector> elements by parsing user-provided information and physical files. Can be called from \code{make_eml()} or as a stand-alone function.
}
\examples{
\dontrun{
# Set working directory
setwd("/Users/me/Documents/data_packages/pkg_260")

# For use with shape_attributes.txt
spatialVector <- create_spatialVector(
  path = "./metadata_templates",
  data.path = "./data_objects")


# For use without a template

my_attributes <- data.frame(
  extname = "my_shape_layer",
  root_dir = "Shapefile Folder",
  description  = "Experimental Plots",
  geoDescription = 'Greater Rio Grande valley area',
  overwrite = TRUE
  )

spatialVector <- create_spatialVector(
  path = "./metadata_templates",
  data.path = "./data_objects",
  vector_attributes = my_attributes)
}

}
