% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/define_factors.R
\name{define_factors}
\alias{define_factors}
\title{Define factors and write to file}
\usage{
define_factors(path)

    Run this function whenever your data contain attributes of class "factor" as listed in 
    \emph{datasetname_datatablename_attributes.txt} files.
}
\arguments{
\item{path}{A path to the dataset working directory containing 
\emph{datasetname_datatablename_attributes.txt}, 
\emph{eml_configuration.R}, and referenced data table files.}
}
\value{
A tab delimited file in the dataset working directory titled 
    \emph{datasetname_datatablename_factors.txt} containing unique values of attributes of class "factor" 
    which is translated and written to EML with \code{make_eml}.
}
\description{
Identify and define data table factors.
}
\details{
This function overwrites any 
    \emph{datasetname_datatablename_factors.txt} files you have created in 
    the dataset working directory. To prevent overwriting of such 
    files, temporarily move them out of the working directory.
    
    This function identifies "factor" class attributes from the file 
    \emph{datasetname_datatablename_attributes.txt} and extracts 
    unique values of these attributes to the table for you to provide definitions for.
    Do not define factors with empty field contents. Delete these rows from this file.
    
    When defining factors with unit values, refer to the standard unit 
    dictionary "name" column. Enter the unit name in the definition column of 
    the factors table. Note these values are case sensitive.
}
\seealso{
\code{\link{import_templates}} to import metadata templates to the 
    dataset working directory.

\code{\link{view_instructions}} for instructions on completing the template 
    files.

\code{\link{extract_geocoverage}} to extract detailed geographic 
    coordinates of sampling sites.

\code{\link{make_eml}} to translate user supplied information into the EML file.
}
