% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{delimiter_infer}
\alias{delimiter_infer}
\title{Infer field delimiter from file name}
\usage{
delimiter_infer(x)
}
\arguments{
\item{x}{(character) File name including path}
}
\value{
(character) Delimiter
}
\description{
Infer field delimiter from file name
}
