% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{detect_delimeter}
\alias{detect_delimeter}
\title{Get field delimiters of input files}
\usage{
detect_delimeter(
  path, 
  data.files, 
  os
)
}
\arguments{
\item{path}{(character) Path to files.}

\item{data.files}{(character) File names.}

\item{os}{(character) Operating system. Valid options are returned from
\code{detect_os}.}
}
\value{
(character) Field delimiters of input files.
    \item{"\\t"}{tab}
    \item{","}{comma}
    \item{";"}{semi-colon}
    \item{"|"}{pipe}
}
\description{
Detect and return field delimiters of input files (tables).
}
