% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{detect_delimeter_2}
\alias{detect_delimeter_2}
\title{Detect field delimiter 2}
\usage{
detect_delimeter_2(data.file, delim.guess)
}
\arguments{
\item{data.file}{(character) Data file name.}

\item{delim.guess}{(character) Delimiter guessed from `detect_delimeter`.}
}
\value{
If ambiguity exists, a manual overide option is presented.
}
\description{
Secondary check on delimeter detection with manual override
}
