% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detect_eol.R
\name{detect_eol}
\alias{detect_eol}
\title{Detect end of line (EOL) character}
\usage{
detect_eol(path = "", file.name = "", os = "")
}
\arguments{
\item{path}{A character string specifying a path to the target file directory.}

\item{file.name}{A character string specifying the target file name.}

\item{os}{A character string specifying the operating system in which this 
function is to be called. Valid options are generated from 
\code{detect_os}.}
}
\value{
A character string representation of the EOL character.
}
\description{
Detect EOL character of input input files and return value
}
