% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detect_non_utf8.R
\name{detect_non_utf8}
\alias{detect_non_utf8}
\title{Detect non-UTF-8 encoded character strings}
\usage{
detect_non_utf8(str = "", method = "")
}
\arguments{
\item{x}{A character vector.
\itemize{
    \item{table} Check character strings of table and report column, 
    row, and value of offending character string.
    row of offending entity.
    \item{text} Check paragraphs of input text and report line number 
    and value of character string.
}}
}
\description{
Identify non-UTF-8 encoded character strings and report location 
    relative to user supplied inputs.
}
