% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eml2eal.R
\name{eml2abstract}
\alias{eml2abstract}
\title{Create abstract template from EML}
\usage{
eml2abstract(eml, path, file.type)
}
\arguments{
\item{eml}{(xml_document xml_node) EML}

\item{path}{(character) Path to write to}

\item{file.type}{(character) File type to write to (see \code{eml2eal()})}
}
\value{
abstract template
}
\description{
Create abstract template from EML
}
\details{
Gets abstract node and parses to template via pandoc.
}
