% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eml2eal.R
\name{eml2additional_info}
\alias{eml2additional_info}
\title{Create additional_info template from EML}
\usage{
eml2additional_info(eml, path, file.type)
}
\arguments{
\item{eml}{(xml_document xml_node) EML}

\item{path}{(character) Path to write to}

\item{file.type}{(character) File type to write to (see \code{eml2eal()})}
}
\value{
additional_info template
}
\description{
Create additional_info template from EML
}
\details{
Gets additionalInfo node and parses to template via pandoc.
}
\keyword{internal}
