% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eml2eal.R
\name{eml2annotations}
\alias{eml2annotations}
\title{Create annotations template from EML}
\usage{
eml2annotations(eml, path)
}
\arguments{
\item{eml}{(xml_document xml_node) EML}

\item{path}{(character) Path to write to}
}
\value{
annotations template
}
\description{
Create annotations template from EML
}
\details{
Gets annotation nodes, listed under subjects or \code{/eml:eml/annotations}, and parse to template.
}
\keyword{internal}
