% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eml2eal.R
\name{eml2categorical_variables}
\alias{eml2categorical_variables}
\title{Create categorical variable template(s) from EML}
\usage{
eml2categorical_variables(eml, path)
}
\arguments{
\item{eml}{(xml_document xml_node) EML}

\item{path}{(character) Path to write to}
}
\value{
categorical variable template(s)
}
\description{
Create categorical variable template(s) from EML
}
\details{
Gets categorical variable nodes and parses to template.
}
