% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eml2eal.R
\name{eml2custom_units}
\alias{eml2custom_units}
\title{Create custom units template from EML}
\usage{
eml2custom_units(eml, path)
}
\arguments{
\item{eml}{(xml_document xml_node) EML}

\item{path}{(character) Path to write to}
}
\value{
custom units template
}
\description{
Create custom units template from EML
}
\details{
Gets unitlist nodes and parses to template.
}
