% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eml2eal.R
\name{eml2eal_losses}
\alias{eml2eal_losses}
\title{Check info lost in the EML to EAL translation}
\usage{
eml2eal_losses(eml, eml.eal)
}
\arguments{
\item{eml}{(character) Full path to original EML file}

\item{eml.eal}{(character) Full path to EML file created by \code{make_eml()}}
}
\value{
(list) Lost values w/xpath names
}
\description{
Check info lost in the EML to EAL translation
}
\examples{
\dontrun{
# EML w/unsupported content
eml <- system.file("eml2eal_test.xml", package = "EMLassemblyline")

# Same EML but after translation through eml2eal()
emleal <- system.file(
  "/examples/pkg_260/eml/edi.260.1.xml", 
  package = "EMLassemblyline")

# Info losses
losses <- eml2eal_losses(eml, emleal)
}

}
