% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eml2eal.R
\name{eml2geographic_coverage}
\alias{eml2geographic_coverage}
\title{Create geographic coverage template from EML}
\usage{
eml2geographic_coverage(eml, path)
}
\arguments{
\item{eml}{(xml_document xml_node) EML}

\item{path}{(character) Path to write to}
}
\value{
geographic coverage template
}
\description{
Create geographic coverage template from EML
}
\details{
Gets geographicCoverage nodes and parses to template.
}
\keyword{internal}
