% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eml2eal.R
\name{eml2intellectual_rights}
\alias{eml2intellectual_rights}
\title{Create intellectual_rights template from EML}
\usage{
eml2intellectual_rights(eml, path, file.type)
}
\arguments{
\item{eml}{(xml_document xml_node) EML}

\item{path}{(character) Path to write to}

\item{file.type}{(character) File type to write to (see \code{eml2eal()})}
}
\value{
intellectual_rights template
}
\description{
Create intellectual_rights template from EML
}
\details{
Gets intellectualRights node and parses to template via pandoc.
}
