% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eml2eal.R
\name{eml2keywords}
\alias{eml2keywords}
\title{Create keywords template from EML}
\usage{
eml2keywords(eml, path)
}
\arguments{
\item{eml}{(xml_document xml_node) EML}

\item{path}{(character) Path to write to}
}
\value{
keywords template
}
\description{
Create keywords template from EML
}
\details{
Gets keywordSet nodes and parses to template.
}
\keyword{internal}
