% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eml2eal.R
\name{eml2make_eml}
\alias{eml2make_eml}
\title{Create \code{make_eml()} function call from EML}
\usage{
eml2make_eml(eml, path)
}
\arguments{
\item{eml}{(xml_document xml_node) EML}

\item{path}{(character) Path to write to}
}
\value{
run_EMLassemblyline_for_pkg.R script containing the \code{make_eml()} function call
}
\description{
Create \code{make_eml()} function call from EML
}
\details{
Gets relevants nodes and parse to script.
}
\keyword{internal}
