% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eml2eal.R
\name{eml2methods}
\alias{eml2methods}
\title{Create methods template from EML}
\usage{
eml2methods(eml, path, file.type)
}
\arguments{
\item{eml}{(xml_document xml_node) EML}

\item{path}{(character) Path to write to}

\item{file.type}{(character) File type to write to (see \code{eml2eal()})}
}
\value{
methods template
}
\description{
Create methods template from EML
}
\details{
Gets methods node, removes provenance methodStep, subsumes sampling and qualityControl, and parses to template via pandoc.
}
