% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eml2eal.R
\name{eml2personnel}
\alias{eml2personnel}
\title{Create personnel template from EML}
\usage{
eml2personnel(eml, path)
}
\arguments{
\item{eml}{(xml_document xml_node) EML}

\item{path}{(character) Path to write to}
}
\value{
personnel template
}
\description{
Create personnel template from EML
}
\details{
Gets personnel and project nodes and parses to template.
}
