% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eml2eal.R
\name{eml2provenance}
\alias{eml2provenance}
\title{Create provenance template from EML}
\usage{
eml2provenance(eml, path)
}
\arguments{
\item{eml}{(xml_document xml_node) EML}

\item{path}{(character) Path to write to}
}
\value{
provenance template
}
\description{
Create provenance template from EML
}
\details{
Gets provenance nodes and parses to template.
}
