% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eml2eal.R
\name{eml2table_attributes}
\alias{eml2table_attributes}
\title{Create table attributes template(s) from EML}
\usage{
eml2table_attributes(eml, path)
}
\arguments{
\item{eml}{(xml_document xml_node) EML}

\item{path}{(character) Path to write to}
}
\value{
table attributes template(s)
}
\description{
Create table attributes template(s) from EML
}
\details{
Gets table attribute nodes and parses to template.
}
