% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eml2eal.R
\name{eml2taxonomic_coverage}
\alias{eml2taxonomic_coverage}
\title{Create taxonomic coverage template from EML}
\usage{
eml2taxonomic_coverage(eml, path)
}
\arguments{
\item{eml}{(xml_document xml_node) EML}

\item{path}{(character) Path to write to}
}
\value{
taxonomic coverage template
}
\description{
Create taxonomic coverage template from EML
}
\details{
Gets taxonomicClassification nodes and parses to template.
}
\keyword{internal}
