% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eml_configuration.R
\name{eml_configuration}
\alias{eml_configuration}
\title{EML configuration}
\usage{
eml_configuration(path)
}
\arguments{
\item{path}{A path to the working directory containing files required by the EMLtools
package. Run the function run_guide() for instructions on how to populate this working
directory with the requisite files.}
}
\value{
A configuration file in the specified working directory containing parameters
for use of functions of the EMLtools package.
}
\description{
A set of user specified parameters for functions of the EMLtools package.
}
\seealso{
\code{\link{run_guide}} for guidance on setting up the working directory

\code{\link{write_attributes}} for writing data attributes

\code{\link{write_factors}} for writing data factors

\code{\link{create_eml}} for creating EML
}
