% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_spatialRaster.R
\name{esri_to_proj4}
\alias{esri_to_proj4}
\title{Get map between ESRI Coordinate System Names and proj4 codes from spatialreference.org}
\usage{
esri_to_proj4()
}
\value{
A data frame wtih columns:
    \item{esri}{ESRI Coordinate System Names}
    \item{proj4}{proj4 codes}
}
\description{
Get map between ESRI Coordinate System Names and proj4 codes from spatialreference.org
}
