% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_geocoverage.R
\name{extract_geocoverage}
\alias{extract_geocoverage}
\title{Write geographic coverage of unique sampling sites to file}
\usage{
extract_geocoverage(path, table.name, lat.col, lon.col, site.col)
}
\arguments{
\item{path}{Path to the directory containing the data table containing latitude, 
longitude, and site name.}

\item{table.name}{Name of the input data table containing geographic 
coverage data.}

\item{lat.col}{Name of latitude column.}

\item{lon.col}{Name of longitude column.}

\item{site.col}{Name of site column.}
}
\value{
A file named datasetname_attributes_draft.xlsx

A file named datasetname_attributes.xlsx

A file named datasetname_attributes.txt
}
\description{
A function to write latitude and longitude of unique sampling
sites to file.
}
