% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{fix_methods}
\alias{fix_methods}
\title{Fix methods section of EML file}
\usage{
fix_methods(eml)
}
\arguments{
\item{eml}{(charcter) Full path to EML document.}
}
\value{
(.xml) The EML file with fixed methods section (see details).
}
\description{
Fix methods section of EML file
}
\details{
This function is the second step in creating an EML methods section from an .md file. Some parsing constraints in \code{EML::write_eml()} require this fix.
}
