% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eml2eal.R
\name{get_anno_context}
\alias{get_anno_context}
\title{Get context for annotations template}
\usage{
get_anno_context(nodeset)
}
\arguments{
\item{nodeset}{(xml_nodeset) Subject nodeset}
}
\value{
(character) Value for context field of annotations template
}
\description{
Get context for annotations template
}
\keyword{internal}
