% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eml2eal.R
\name{get_anno_element}
\alias{get_anno_element}
\title{Get element for annotations template}
\usage{
get_anno_element(nodeset)
}
\arguments{
\item{nodeset}{(xml_nodeset) Subject nodeset}
}
\value{
(character) Value for element field of annotations template
}
\description{
Get element for annotations template
}
