% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eml2eal.R
\name{get_anno_id}
\alias{get_anno_id}
\title{Get id for annotations template}
\usage{
get_anno_id(nodeset)
}
\arguments{
\item{nodeset}{(xml_nodeset) Subject nodeset}
}
\value{
(character) Value for id field of annotations template
}
\description{
Get id for annotations template
}
\keyword{internal}
