% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eml2eal.R
\name{get_anno_subject}
\alias{get_anno_subject}
\title{Get subject for annotations template}
\usage{
get_anno_subject(nodeset)
}
\arguments{
\item{nodeset}{(xml_nodeset) Subject nodeset}
}
\value{
(character) Value for subject field of annotations template
}
\description{
Get subject for annotations template
}
