% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eml2eal.R
\name{get_class}
\alias{get_class}
\title{Get class of table attribute}
\usage{
get_class(nodeset)
}
\arguments{
\item{nodeset}{(xml_nodeset) Attribute nodeset at \code{/eml:eml/dataset/dataTable[x]/attributeList/attribute[y]}}
}
\value{
(character) numeric, character, categorical, or Date
}
\description{
Get class of table attribute
}
\details{
Wrapper to \code{is_num(nodeset)}, \code{is_char(nodeset)}, \code{is_catvar(nodeset)}, \code{is_date(nodeset)}
}
