% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eml2eal.R
\name{get_misscode}
\alias{get_misscode}
\title{Get missing value code}
\usage{
get_misscode(nodeset)
}
\arguments{
\item{nodeset}{(xml_nodeset) Attribute nodeset at \code{/eml:eml/dataset/dataTable[x]/attributeList/attribute[y]}}
}
\value{
(character) Missing value code
}
\description{
Get missing value code
}
\details{
Get from \code{./missingValueCode/code}. Use first and warn if > 2.
}
