% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eml2eal.R
\name{get_misscodedef}
\alias{get_misscodedef}
\title{Get missing value code definition}
\usage{
get_misscodedef(nodeset)
}
\arguments{
\item{nodeset}{(xml_nodeset) Attribute nodeset at \code{/eml:eml/dataset/dataTable[x]/attributeList/attribute[y]}}
}
\value{
(character) Missing value code definition
}
\description{
Get missing value code definition
}
\details{
Get from \code{./attribute/missingValueCode/codeExplanation}. Use first and warn if > 2.
}
\keyword{internal}
