% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eml2eal.R
\name{get_parallel_nodes}
\alias{get_parallel_nodes}
\title{Get parallel node values}
\usage{
get_parallel_nodes(eml, xpath)
}
\arguments{
\item{eml}{(xml_document xml_node) EML}

\item{xpath}{(character) xpath to node}
}
\value{
(character) "" if node is missing from one nodeset but not the other
}
\description{
Get parallel node values
}
\details{
Works for dataTable and otherEntity, not others
}
\keyword{internal}
