% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eml2eal.R
\name{get_proj}
\alias{get_proj}
\title{Get project information}
\usage{
get_proj(eml)
}
\arguments{
\item{eml}{(xml_document xml_node) EML}
}
\value{
(character) givenName[1], givenName[2] (i.e. middleInitial), surName, organizationName, electronicMailAddress, userId, role, title (i.e. projectTitle), funding (i.e. fundingAgency + fundingNumber)
}
\description{
Get project information
}
\details{
Get from \code{/eml:eml/dataset/project} or \code{/eml:eml/dataset/project/relatedProject}.
}
\keyword{internal}
