% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eml2eal.R
\name{get_reference}
\alias{get_reference}
\title{Get referenced node}
\usage{
get_reference(nodeset)
}
\arguments{
\item{nodeset}{(xml_nodeset) Any nodeset}
}
\value{
(xml_nodeset) nodeset with matching id attribute
}
\description{
Get referenced node
}
