% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eml2eal.R
\name{get_resparty}
\alias{get_resparty}
\title{Get responsible party}
\usage{
get_resparty(nodeset)
}
\arguments{
\item{nodeset}{(xml_nodeset) ResponsibleParty nodeset}
}
\value{
(character) givenName[1], givenName[2] (i.e. middleInitial), surName, organizationName, electronicMailAddress, userId, role
}
\description{
Get responsible party
}
\details{
Get from \code{/eml:eml/dataset/creator}, \code{/eml:eml/dataset/contact}, \code{/eml:eml/dataset/associatedParty}.
}
\keyword{internal}
