% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eml2eal.R
\name{get_unit}
\alias{get_unit}
\title{Get unit}
\usage{
get_unit(nodeset)
}
\arguments{
\item{nodeset}{(xml_nodeset) Attribute nodeset at \code{/eml:eml/dataset/dataTable[x]/attributeList/attribute[y]}}
}
\value{
(character) unit
}
\description{
Get unit
}
\details{
Get unit from \code{./unit/standardUnit} or \code{./attribute/measurementScale/ratio/unit/customUnit}
}
