% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eml2eal.R
\name{get_xpaths_and_vals}
\alias{get_xpaths_and_vals}
\title{Get all xpaths and values in a nodeset}
\usage{
get_xpaths_and_vals(nodeset)
}
\arguments{
\item{nodeset}{(xml_nodeset/xml_node) A nodeset}
}
\value{
(list) All xpaths and values in \code{nodeset}, where xpaths are stored as value names
}
\description{
Get all xpaths and values in a nodeset
}
