% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eml2eal.R
\name{has_parent}
\alias{has_parent}
\title{Has parent node}
\usage{
has_parent(nodeset, parent)
}
\arguments{
\item{nodeset}{(xml_nodeset) A nodeset}

\item{parent}{(character) Name of parent to search for}
}
\value{
(logical) TRUE if \code{parent} is in the list of parents
}
\description{
Has parent node
}
