% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{init_attributes}
\alias{init_attributes}
\title{Initialize an empty attributes template}
\usage{
init_attributes(nrows = 0)
}
\arguments{
\item{nrows}{(numeric) Number of rows to initialize the data frame with.}
}
\value{
(data.frame) The attributes template.
}
\description{
Initialize an empty attributes template
}
\examples{
\dontrun{
# Default settings return an empty data frame with column names of the 
# attributes template.
df = init_attributes()
nrow(df)
colnames(df)

# Control the number of rows with the nrows parameter
df = init_attributes(nrows = 3)
nrow(df)
}

}
\keyword{internal}
