% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{init_catvars}
\alias{init_catvars}
\title{Initialize an empty categorical variables template}
\usage{
init_catvars(nrows = 0)
}
\arguments{
\item{nrows}{(numeric) Number of rows to initialize the data frame with.}
}
\value{
(data.frame) The categorical variables template.
}
\description{
Initialize an empty categorical variables template
}
\examples{
\dontrun{
# Default settings return an empty data frame with column names of the 
# categorical variables template.
df = init_catvars()
nrow(df)
colnames(df)

# Control the number of rows with the nrows parameter
df = init_catvars(nrows = 3)
nrow(df)
}

}
\keyword{internal}
