% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{init_entities}
\alias{init_entities}
\title{Initialize an empty entities template}
\usage{
init_entities(nrows = 0)
}
\arguments{
\item{nrows}{(numeric) Number of rows to initialize the data frame with.}
}
\value{
(data.frame) The entities template.
}
\description{
Initialize an empty entities template
}
\examples{
\dontrun{
# Default settings return an empty data frame with column names of the 
# entities template.
df = init_entities()
nrow(df)
colnames(df)

# Control the number of rows with the nrows parameter
df = init_entities(nrows = 3)
nrow(df)
}

}
\keyword{internal}
