% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eml2eal.R
\name{is_catvar}
\alias{is_catvar}
\title{Is attribute a categorical variable?}
\usage{
is_catvar(nodeset)
}
\arguments{
\item{nodeset}{(xml_nodeset) Attribute nodeset at \code{/eml:eml/dataset/dataTable[x]/attributeList/attribute[y]}}
}
\value{
(logical) TRUE if nodeset has \code{./measurementScale/nominal/nonNumericDomain/enumeratedDomain} or \code{./measurementScale/ordinal/nonNumericDomain/enumeratedDomain}
}
\description{
Is attribute a categorical variable?
}
