% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eml2eal.R
\name{is_date}
\alias{is_date}
\title{Is attribute a Date?}
\usage{
is_date(nodeset)
}
\arguments{
\item{nodeset}{(xml_nodeset) Attribute nodeset at \code{/eml:eml/dataset/dataTable[x]/attributeList/attribute[y]}}
}
\value{
(logical) TRUE if nodeset has \code{./measurementScale/dateTime}
}
\description{
Is attribute a Date?
}
