% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eml2eal.R
\name{is_num}
\alias{is_num}
\title{Is attribute a numeric variable?}
\usage{
is_num(nodeset)
}
\arguments{
\item{nodeset}{(xml_nodeset) Attribute nodeset at \code{/eml:eml/dataset/dataTable[x]/attributeList/attribute[y]}}
}
\value{
(logical) TRUE if nodeset has \code{./attribute/measurementScale/interval} or \code{./attribute/measurementScale/ratio}
}
\description{
Is attribute a numeric variable?
}
\keyword{internal}
