% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{is_template}
\alias{is_template}
\title{Is a metadata template file}
\usage{
is_template(files)
}
\arguments{
\item{files}{(character) Vector of file names. Can be full file paths or
or base names.}
}
\value{
(logical) TRUE where \code{files} is recognized as a metadata 
template.
}
\description{
Check which files are metadata templates and which are not.
}
\examples{
\dontrun{
testdir <- paste0(tempdir(), "/pkg")
dir_files <- copy_test_package(testdir)
i <- is_template(dir_files)
i
unlink(testdir, recursive = TRUE, force = TRUE)
}

}
\keyword{internal}
