% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{list_attribute_templates}
\alias{list_attribute_templates}
\title{List attribute templates in a directory}
\usage{
list_attribute_templates(path)
}
\arguments{
\item{path}{(character) Path to the metadata template directory.}
}
\value{
(character) File names of attribute templates.
}
\description{
List attribute templates in a directory
}
\examples{
\dontrun{

}

}
\keyword{internal}
