% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lter_id.R
\name{lter_id}
\alias{lter_id}
\title{lter_id}
\usage{
lter_id(x)
}
\arguments{
\item{x}{(character) A valid term in the LTER Controlled Vocabulary.}
}
\value{
(numeric) The identification number for a LTER Controlled Vocabulary 
    term.
}
\description{
Get the identification number for a valid term in the LTER Controlled 
    Vocabulary.
}
