% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lter_scope.R
\name{lter_scope}
\alias{lter_scope}
\title{lter_scope}
\usage{
lter_scope(id)
}
\arguments{
\item{id}{(numeric) An identification number of a valid term in the LTER 
Controlled Vocabulary.}
}
\value{
(character) The scope description for a LTER Controlled Vocabulary 
    term. Note, not all terms have descriptions.
}
\description{
Get the scope description for a term in the LTER Controlled Vocabulary.
}
