% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lter_term.R
\name{lter_term}
\alias{lter_term}
\title{lter_term}
\usage{
lter_term(x, messages = FALSE)
}
\arguments{
\item{x}{(character) A term to search for.}

\item{messages}{(logical) Display diagnostic messages, e.g. alternative spelling options.}
}
\value{
TRUE - If the searched term is found.
    FALSE- If the searched term was not found.
    messages - Alternative spelling and near misses for searched term.
}
\description{
Search for a term in the LTER Controlled Vocabulary (Long Term 
    Ecological Research Network).
}
