% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_arguments.R
\name{make_arguments}
\alias{make_arguments}
\title{Initialize list of arguments for EMLassemblyline functions}
\usage{
make_arguments(
      path = NULL,
      data.path = NULL,
      data.table = NULL,
      other.entity = NULL,
      sep = NULL
    )
}
\arguments{
\item{path}{(character) Path to metadata templates.}

\item{data.path}{(character) Path to data tables and other entities.}

\item{data.table}{(character) Data table name. If more than one, then supply as a vector 
of character strings (e.g. 
\code{data.table = c('concentrations.csv', 'characteristics.csv')}).}

\item{other.entity}{(character) Other entity name. If more than one, then supply as a vector 
of character strings (e.g. 
\code{other.entity = c('ancillary_data.zip', 'quality_control.R')}).}

\item{sep}{(character) Data table field delimiter. Use this argument if 
\code{make_arguments} fails to automatically identify field delimiters of 
\code{data.table}.}
}
\description{
Initialize the list of arguments for EMLassemblyline functions including
    \code{x}, an argument containing metadata template and data file 
    content. This approach enables programatic interfacing of upstream 
    metadata sources (e.g. 
    \href{https://github.com/lter/LTER-core-metabase}{LTER-core-metabase}) 
    to \code{EMLassemblyline} functions.
}
\note{
The full named list structure is returned irrespective of NULL inputs,
    so content can be added downstream. If valid inputs are supplied, then
    the respective list structures are populated. Pass the output of this 
    function to \code{make_template_core} to initilize (or add) core 
    template data frames, text objects, or XML objects.
}
