% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{name_attribute_templates}
\alias{name_attribute_templates}
\title{Create attribute template names from file names}
\usage{
name_attribute_templates(files)
}
\arguments{
\item{files}{(character vector) Names of data object files, including file 
extensions.}
}
\value{
(character vector) Names of attribute templates.
}
\description{
Create attribute template names from file names
}
\examples{
f <- c("decomp.csv", "nitrogen.csv")
name_attribute_templates(f)

}
\keyword{internal}
