% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{name_attributes_templates}
\alias{name_attributes_templates}
\title{Create attributes template file names from data object file names}
\usage{
name_attributes_templates(files)
}
\arguments{
\item{files}{(character vector) Names of data object files, including file 
extensions.}
}
\value{
(character vector) Names of attribute templates.
}
\description{
Create attributes template file names from data object file names
}
\examples{
\dontrun{
f <- c("decomp.csv", "nitrogen.csv")
name_attributes_templates(f)
}

}
\keyword{internal}
