% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{name_catvars_templates}
\alias{name_catvars_templates}
\title{Create categorical variables template file names from data object file names}
\usage{
name_catvars_templates(files)
}
\arguments{
\item{files}{(character vector) Names of data object files, including file 
extensions.}
}
\value{
(character vector) Names of categorical variables templates.
}
\description{
Create categorical variables template file names from data object file names
}
\examples{
\dontrun{
f <- c("decomp.csv", "nitrogen.csv")
name_catvars_templates(f)
}

}
\keyword{internal}
