% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{name_data_objects}
\alias{name_data_objects}
\title{Create data object file names from attributes template file names}
\usage{
name_data_objects(files, data.path)
}
\arguments{
\item{files}{(character) File names of attributes templates, including the 
file extension.}

\item{data.path}{(character) Path to the data directory.}
}
\value{
(character) File names of the data objects corresponding with 
\code{files}.
}
\description{
Create data object file names from attributes template file names
}
\examples{
\dontrun{
testdir <- paste0(tempdir(), "/pkg")
dir_files <- copy_test_package(testdir)
attr_files <- c("attributes_decomp.txt", "attributes_nitrogen.txt")
name_data_objects(attr_files, testdir)
unlink(testdir, recursive = TRUE, force = TRUE)
}

}
\keyword{internal}
