% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eml2eal.R
\name{nodeset2txt}
\alias{nodeset2txt}
\title{Convert nodeset to EAL text type template}
\usage{
nodeset2txt(nodeset, file.type, path)
}
\arguments{
\item{nodeset}{(xml_nodeset) Nodeset to convert}

\item{file.type}{(character) File type to write to (see \code{eml2eal()})}

\item{path}{(character) Path to which outputs will be written}
}
\value{
abstract, methods, additional_info, or intellectual_rights
}
\description{
Convert nodeset to EAL text type template
}
\details{
Converts EML text type \code{nodeset} (i.e. abstract, methods, additionalInfo, intellectualRights) to EAL \code{file.type} via pandoc
}
