% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{read_data_objects}
\alias{read_data_objects}
\title{Read data objects}
\usage{
read_data_objects(data.path, data.objects)
}
\arguments{
\item{data.path}{(character) Path to the data directory.}

\item{data.objects}{(character) File names of data objects.}
}
\value{
(list) Named list of parsed data objects with the files:
\itemize{
\item{content: The data object, in it's R representation. This value will be
\code{NA} if the object could not be read.}
\item{mime_type: The data object's MIME type.}
\item{file_path: The data object's file path.}
}
}
\description{
Attempts to read data objects through a series of parsers informed by the
data object mime type.
}
\examples{
\dontrun{
test_dir <- paste0(tempdir(), "/pkg")
pkg_files <- copy_test_package(test_dir)
d <- read_data_objects(
  data.path = test_dir,
  data_objects = c(
    "ancillary_data.zip",
    "decomp.csv",
    "nitrogen.csv",
    "processing_and_analysis.R"
  )
)
View(d)
unlink(test_dir, recursive = TRUE, force = TRUE)
}

}
\keyword{internal}
