% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_files.R
\name{read_files}
\alias{read_files}
\title{Read templates and data files into list (alternative input to 
`EMLassemblyline` functions)}
\usage{
read_file_list(path, data.path = path, data.table = NULL, 
other.entity = NULL)
}
\arguments{
\item{path}{(character) Path to where the template(s) will be imported.}

\item{data.path}{(character) Path to where the data tables are stored.}

\item{data.table}{(character) Data table name(s). If more than one, then supply as a 
vector of character strings (e.g. 
`data.table = c("concentrations.csv", "characteristics.csv")`).}

\item{other.entity}{(character) Name(s) of `other.entity`(s) in this dataset. Use
`other.entity` for all non-`data.table` files. `other.entity`(s) should 
be stored at `data.path`.}
}
\value{
(named list) Named list containing these elements
    \itemize{
        \item{`name` Name of file}
        \item{`path` Where the file was read from}
        \item{`contents` Contents of the file (e.g. data.frame, XML, etc.)}
    }
}
\description{
This function is under development.
}
\details{
Salient details of this function
}
