% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{read_template_characteristics}
\alias{read_template_characteristics}
\title{Read template characteristics table}
\usage{
read_template_characteristics()
}
\value{
(data.frame) The template characteristics table.
}
\description{
The template characteristics table contains information to aid in template 
reading and parsing operations.
}
\examples{
\dontrun{
tmplt_chars <- read_template_characteristics
tmplt_chars
}

}
\keyword{internal}
