% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resolve_terms.R
\name{resolve_terms}
\alias{resolve_terms}
\title{resolve_terms}
\usage{
resolve_terms(x, cv)
}
\arguments{
\item{x}{(character) Term(s) to resolve to a controlled vocabulary. Can be a 
vector of terms.}

\item{cv}{(character) A controlled vocabulary to search. Valid options are:
\itemize{
    \item{lter} - The LTER Controlled Vocabulary (http://vocab.lternet.edu/vocab/vocab/index.php)
}}

\item{messages}{(logical) Display diagnostic messages, e.g. alternative spelling options.}
}
\value{
(character) Controlled vocabulary names corresponding to successfully
    resolved terms.
}
\description{
Resolve terms to a controlled vocabulary.
}
