% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eml2eal.R
\name{rm_pred}
\alias{rm_pred}
\title{Remove xpath predicates}
\usage{
rm_pred(xpath)
}
\arguments{
\item{xpath}{(character) xpath}
}
\value{
(character) \code{xpath} with predicates removed
}
\description{
Remove xpath predicates
}
