% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{set_methods_md}
\alias{set_methods_md}
\title{Create methods node from .md file}
\usage{
set_methods_md(methods_file)
}
\arguments{
\item{methods_file}{(character) Full path to methods.md file}
}
\value{
(list) EML methodStep node
}
\description{
Create methods node from .md file
}
\details{
Text and LaTeX equations in \code{methods_file} are parsed into /eml/dataset/methodStep/description/markdown nodes. A second step applied by \code{fix_methods()} on the resultant EML file corrects the node names. Without this second step, the EML file will fail on schema validation.
}
\note{
LaTex equations must be wrapped in $$ (e.g. $$<my_equation>$$), otherwise they won't be parsed correctly from the text.
}
