% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/template_categorical_variables.R
\name{table_to_attribute_template}
\alias{table_to_attribute_template}
\title{Convert data table file name to the corresponding attributes file name}
\usage{
table_to_attribute_template(data.table)
}
\arguments{
\item{data.table}{(character) File name of data table, including file extension}
}
\value{
(character) Table attributes file name
}
\description{
Convert data table file name to the corresponding attributes file name
}
\keyword{internal}
