% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/template_annotations.R
\name{template_annotations}
\alias{template_annotations}
\title{Create annotations template}
\usage{
template_annotations(
      path,
      write.file = TRUE,
      x = NULL
    )
}
\arguments{
\item{path}{(character) Path to the metadata template directory.}

\item{write.file}{(logical; optional) Whether to write the template file.}

\item{x}{(named list; optional) Alternative input to 
\code{template_annotations()}. Use \code{template_arguments()} 
to create \code{x}.}
}
\value{
\strong{annotations.txt} The tab delimited annotations template. This 
    file is written to \code{path} unless using \code{x}, in which case the 
    template is added to \strong{/x/templates/annotations.txt}.
}
\description{
Use this function to extract annotatable elements from completed 
    metadata templates.
    \href{https://ediorg.github.io/EMLassemblyline/articles/edit_metadata_templates.html}{Instructions for editing this template.}
}
\details{
An existing annotations template will not be overwritten by subsequent 
    calls to \code{template_annotations()}.
}
\examples{
# Initialize data package directory for template_annotations()
file.copy(
 from = system.file('/examples/pkg_260', package = 'EMLassemblyline'),
 to = tempdir(),
 recursive = TRUE
)

# Set working directory
setwd(paste0(tempdir(), '/pkg_260'))

# View directory contents (NOTE: annotations.txt dosn't exist)
dir('./metadata_templates')

# Template annotations
template_annotations(
  path = './metadata_templates'
)

# View directory contents (NOTE: annotations.txt exists)
dir('./metadata_templates')

# View template contents
df <- data.table::fread(./metadata_templates/annotations.txt)
df

# Rerunning template_annotations() does not overwrite files
template_annotations(
  path = './metadata_templates'
)

# Clean up
unlink('.', recursive = TRUE)
    
}
