% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/template_categorical_variables.R
\name{template_categorical_variables}
\alias{template_categorical_variables}
\title{Describe categorical variables of a data table}
\usage{
template_categorical_variables(path, data.path = path, write.file = TRUE)
}
\arguments{
\item{path}{(character) Path to the metadata template directory.}

\item{data.path}{(character) Path to the data directory.}

\item{write.file}{(logical; optional) Whether to write the template file.}
}
\value{
\item{catvars_tablename.txt}{The tab delimited categorical variable template, where "tablename" is the table name from which the variables were extracted. This file is written to \code{path}}.
\item{list of data frames}{A list of data frames. One for each categorical variables template.}
}
\description{
Describes categorical variables of a data table. Use if any columns are classified as categorical in table attributes template.
}
\details{
\code{template_categorical_variables()} knows which columns of a table
    are \code{categorical} based on their definition under the \code{class} 
    column of the attributes_*.txt template.
    
    An existing categorical variables template will not be overwritten 
    by subsequent calls to \code{template_categorical_variables()}.
    
    Character encoding of metadata extracted directly from the tables are 
    converted to UTF-8 via \code{enc2utf8()}.
}
\examples{
\dontrun{
# Set working directory
setwd("/Users/me/Documents/data_packages/pkg_260")

# For tables containing categorical variables as classified in the table attributes template
template_categorical_variables(
  path = "./metadata_templates",
  data.path = "./data_objects")
}

}
