% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/template_core_metadata.R
\name{template_core_metadata}
\alias{template_core_metadata}
\title{Create core metadata templates}
\usage{
template_core_metadata(
      path,
      license,
      write.file = TRUE,
      x = NULL
    )
}
\arguments{
\item{path}{(character) Path to the metadata template directory.}

\item{license}{(character) License to publicly release the data package under. Use
\href{https://creativecommons.org/publicdomain/zero/1.0/}{"CC0"} or 
\href{https://creativecommons.org/licenses/by/4.0/}{"CCBY"}.}

\item{x}{(named list; optional) Alternative input to 
\code{template_core_metadatta()}. Use \code{template_arguments()} 
to create \code{x}.}

\item{write.file}{(logical; optional) Whether to write the template files.}
}
\value{
\itemize{
        \item{\strong{abstract.txt} The abstract template.}
        \item{\strong{additional_info.txt} The template for miscellaneous
        information.}
        \item{\strong{intellectual_rights.txt} The intellectual rights 
        license with the text of CC0 or CCBY.}
        \item{\strong{keywords.txt} The tab delimited keywords template.}
        \item{\strong{methods.txt} The methods template.}
        \item{\strong{personnel.txt} The tab delimited personnel template 
        for information on persons and funding involved in the creation of 
        the data package.}
        \item{These files is written to \code{path} unless using \code{x},
        in which case the templates are added to 
        \strong{/x/templates}.}
    }
}
\description{
Use this function to create the core metadata templates required by 
    all data packages. Most templates require user supplied information.
    \href{https://ediorg.github.io/EMLassemblyline/articles/edit_metadata_templates.html}{Instructions for editing these templates.}
}
\details{
Existing templates will not be overwritten by \code{template_core_metadata()}.
}
\examples{
# Initialize data package directory for template_core_metadata()
file.copy(
  from = system.file('/examples/pkg_250', package = 'EMLassemblyline'),
  to = tempdir(),
  recursive = TRUE
)

# Set working directory
setwd(paste0(tempdir(), '/pkg_250'))

# View directory contents (NOTE: Directory is empty)
dir('./metadata_templates')

# Template core metadata
template_core_metadata(
  path = './metadata_templates',
  license = 'CC0'
)

# View directory contents (NOTE: Templates exist)
dir('./metadata_templates')

# Rerunning template_core_metadata() does not overwrite files
template_core_metadata(
  path = './metadata_templates',
  license = 'CC0'
)

# Clean up
unlink('.', recursive = TRUE)
    
}
