% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/template_directories.R
\name{template_directories}
\alias{template_directories}
\title{Create directories for \code{EMLassemblyline} inputs and outputs}
\usage{
template_directories(path, dir.name)
}
\arguments{
\item{path}{(character) Path to where the directory structure will be created.}

\item{dir.name}{(character) Name of parent directory. Subdirectories have predefined 
names.}
}
\description{
Create a commonly used directory structure for \code{EMLassemblyline} 
    inputs and outputs. You're welcome to organize content in any way you 
    choose, this is just one approach.
}
\note{
Existing directories named with \code{dir.name} at \code{path} will not 
    be overwritten.
}
