% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/template_entities.R
\name{template_entities}
\alias{template_entities}
\title{Describe entities of data objects}
\usage{
template_entities(
  path,
  data.path = path,
  data.objects = NULL,
  empty = FALSE,
  write.file = TRUE
)
}
\arguments{
\item{path}{(character) Path to the metadata template directory.}

\item{data.path}{(character) Path to the data directory.}

\item{data.objects}{(character) File names of data objects. If \code{NULL},
data objects with an attribute template will be used.}

\item{empty}{(logical) Whether to create an empty template.}

\item{write.file}{(logical) Whether to write the template to file. If
\code{FALSE}, a data frames will be returned.}
}
\value{
If \code{write.file = TRUE}, tab delimited file is written to \code{path}. If
\code{write.file = FALSE} a data frames is returned. Columns of
this template:
\itemize{
\item{objectName: name of the described object}
\item{variable: described metadata}
\item{value: value for the metadata set in the \code{variable} field for the 
object \code{objectName}}
}
}
\description{
Use this function to describe data entity metadata. This file will be used
in the make_eml function to set the values defined in this template. 
Currently, only the entityType metadata variable for otherEntity 
is supported.
}
\note{
Currently, only the entityType metadata variable for otherEntity 
is supported.
}
\examples{
\dontrun{
# Create a temporary directory with files for this example
testdir <- paste0(tempdir(), "/pkg")
pkg_files <- copy_test_package(testdir)

# Return template as data frames
tmplt <- template_entities(path = testdir, write.file = FALSE)

# Return template as file
template_entities(path = testdir, write.file = TRUE)

# With the data.objects argument
template_entities(
  path = testdir, 
  data.objects = c("ancillary_data.zip", "processing_and_analysis.R"), 
  write.file = TRUE
)

# Clean up files of this example
unlink(testdir, force = TRUE)#' }

}
