% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/template_geographic_coverage.R
\name{template_geographic_coverage}
\alias{template_geographic_coverage}
\title{Create geographic coverage template}
\usage{
template_geographic_coverage(
      path, 
      data.path = path, 
      data.table, 
      lat.col, 
      lon.col, 
      site.col, 
      write.file = TRUE,
      x = NULL
    )
}
\arguments{
\item{path}{(character) Path to the directory where the geographic coverage 
template will be written.}

\item{data.path}{(character) Path to the directory containing \code{data.table}.}

\item{data.table}{(character) Name of table containing geographic coordinates 
represented in decimal degrees, where latitudes south of the equator 
and longitudes west of the prime meridian are negative.}

\item{lat.col}{(character) Name of latitude column.}

\item{lon.col}{(character) Name of longitude column.}

\item{site.col}{(character) Name of site column, where site is the name of the location
specified by \code{lat.col} and \code{lon.col}.}

\item{x}{(named list; optional) Alternative input to 
\code{EMLassemblyline} functions. Use \code{template_arguments()} to 
create \code{x}.}

\item{write.file}{(logical; optional) Whether to write the geographic coverage template 
to \code{path}.}
}
\value{
\itemize{
        \item{\strong{geographic_coverage.txt} The geographic 
        coverage template. A tab delimited table.}
        \item{If using \code{x}, then the geographic coverage template is 
        added to \strong{/x/templates}.}
    }
}
\description{
Create template for describing the geographic coverage of a data package.
    The content of this template is automatically extracted from a data 
    table containing geographical coordinates.
}
\details{
An existing geographic coverage template will not be overwritten by 
    subsequent calls to \code{extract_geocoverage()}.
}
\examples{
# Set working directory
setwd(tempdir())

# Create data package directory "edi_255"
file.copy(
  from = system.file('/examples/edi_255', package = 'EMLassemblyline'),
  to = '.',
  recursive = TRUE
)

# View directory contents (NOTE: geographic_coverage.txt doesn't exist)
dir('./edi_255/metadata_templates')

# Template geographic coverage
template_geographic_coverage(
  path = './edi_255/metadata_templates',
  data.path = './edi_255/data_objects',
  data.table = 'nitrogen.csv',
  site.col = 'site_name',
  lat.col = 'site_lat',
  lon.col = 'site_lon'
)

# View directory contents (NOTE: geographic_coverage.txt exists)
dir('./edi_255/metadata_templates')

# Rerunning template_geographic_coverage() does not overwrite files
template_geographic_coverage(
  path = './edi_255/metadata_templates',
  data.path = './edi_255/data_objects',
  data.table = 'nitrogen.csv',
  site.col = 'site_name',
  lat.col = 'site_lat',
  lon.col = 'site_lon'
)

}
