% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/template_netcdf_attributes.R
\name{template_netcdf_attributes}
\alias{template_netcdf_attributes}
\title{Describe NetCDf files}
\usage{
template_netcdf_attributes(
  path,
  data.path = path,
  netcdf = NULL,
  write.file = TRUE
)
}
\arguments{
\item{path}{(character) Path to the metadata template directory.}

\item{data.path}{(character) Path to the data directory.}

\item{netcdf}{(character) file name. If more than one, then supply as a 
vector of character strings (e.g. 
\code{netcdf = c('netcdf_1.nc', 'netcdf_2.nc')}).}

\item{write.file}{(logical; optional) Whether to write the template file.}
}
\value{
\item{attributes_*}{Columns:
    \itemize{
    \item{attributeName: variable/dimension name}
    \item{attributeDefinition: variable/dimension definition}
    \item{class: variable/dimension class. Valid options are: "numeric" (Numeric variable), "categorical" (Categorical variable, i.e. nominal), "character" (Free text character strings, e.g. notes), "Date" (Date and time variable)}
    \item{unit: variable/dimension unit. Required for numeric classes. Select from EML's standard unit dictionary, accessible with \code{view_unit_dictionary()}. Use values in the "id" column. If not found, then define as a custom unit (see custom_units.txt).}
    \item{dateTimeFormatString: Format string. Required for Date classes. Valid format string components are: "Y" (Year), "M" (Month), "D" (Day), "h" (Hour), "m" (Minute), "s" (Second), Common separators of format string components (e.g. "-" "/" "\" ":"") are supported.}
    \item{missingValueCode: Missing value code. Required for variables containing a missing value code).}
    \item{missingValueCodeExplanation: Definition of missing value code.}
    }
}
\item{custom_units}{Describes non-standard units used in a NetCDF attribute template. Columns:
    \itemize{
    \item{id: Unit name listed in the unit column of the table attributes template (e.g. feetPerSecond)}
    \item{unitType: Unit type (e.g. velocity)}
    \item{parentSI: SI equivalent (e.g. metersPerSecond)}
    \item{multiplierToSI: Multiplier to SI equivalent (e.g. 0.3048)}
    \item{Abbreviation: Abbreviation (e.g. ft/s)}
    }
}
}
\description{
Describes variables and dimensions of a NetCDF file
 (classes, units, datetime formats, missing value codes).
}
\details{
Character encoding of metadata extracted directly from the tables are 
    converted to UTF-8 via \code{enc2utf8()}.
}
\examples{
\dontrun{
# Set working directory
setwd("/Users/me/Documents/data_packages/pkg_260")

template_netcdf_attributes(
  path = './metadata_templates',
  data.path = './data_objects',
  netcdf = 'netcdf_test_file.nc')
}
    
}
