% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/template_netcdf_catvars.R
\name{template_netcdf_catvars}
\alias{template_netcdf_catvars}
\title{Describe categorical variables of a NetCDF file}
\usage{
template_netcdf_catvars(
  path = NULL,
  data.path = path,
  netcdf = NULL,
  write.file = TRUE
)
}
\arguments{
\item{path}{(character) Path to the metadata template directory.}

\item{data.path}{(character) Path to the data directory.}

\item{netcdf}{(character) File names of NetCDF. If more than one, then supply as 
a vector.}

\item{write.file}{(logical; optional) Whether to write the template file.}
}
\value{
\item{catvars_*.txt}{Columns:
    \itemize{
    \item{attributeName: Variable or dimension name}
    \item{code: Categorical variable}
    \item{definition: Definition of categorical variable}
    }
}
}
\description{
Describes the categorical variable represented in a NetCDF file.
}
\examples{
\dontrun{
# Set working directory
setwd("/Users/me/Documents/data_packages/pkg_260")

# For files containing categorical variables as classified in the netcdf attributes template
template_netcdf_catvars(
  path = "./metadata_templates",
  data.path = "./data_objects",
  netcdf = setNames(
    c('netcdf_test_file.nc'),
    c('attributes_netcdf_test_file.txt')
  ))
}

}
