% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/template_other_entity.R
\name{template_other_entity_attributes}
\alias{template_other_entity_attributes}
\title{Create other entity attributes template}
\usage{
template_other_entity_attributes(
      path,
      data.path = path,
      other.entity = NULL,
      write.file = TRUE,
      x = NULL
    )
}
\arguments{
\item{path}{(character) Path to the metadata template directory.}

\item{data.path}{(character) Path to the data directory.}

\item{other.entity}{(character) Other entity name. If more than one, then supply as a 
vector of character strings (e.g. 
\code{other.entity = c('maps.zip', 'analysis_script.R')}).}

\item{write.file}{(logical; optional) Whether to write the template file.}

\item{x}{(named list; optional) Alternative input to 
\code{template_table_attributes()}. Use \code{template_arguments()} 
to create \code{x}.}
}
\value{
\itemize{
        \item{\strong{attributes_dataset.txt} The tab delimited dataset 
        attributes template updated with other entity IDs, file names, and 
        additional rows for specifying other entity names and descriptions.
        This file is written to \code{path} unless using \code{x}, 
        in which case the updated template is added to 
        \strong{/x/templates/attributes_*.txt}.}
        \item{\strong{attributes_other_entity.txt} (coming soon!) The tab 
        delimited other entity template for defining other entity attributes,
        written to \code{path} unless using \code{x}, in which case the 
        template is added to 
        \strong{/x/templates/attributes_other_entity.txt}.}
    }
}
\description{
Implementation of this function is not yet complete.
    Use this function to extract attributes from data objects classified as
    "other entity" (i.e. data that are not tables, spatial raster, or
    spatial vector types) and return for user supplied definitions.
}
\details{
An existing attributes template will not be overwritten by subsequent 
    calls to \code{template_other_entity_attributes()}.
}
\examples{
    
}
