% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/template_entities.R
\name{template_otherentity_entities}
\alias{template_otherentity_entities}
\title{Describe entities of otherEntity data objects}
\usage{
template_otherentity_entities(data.object, empty)
}
\arguments{
\item{data.object}{(list) data object to describe}

\item{empty}{(logical) Whether to create an empty template.}
}
\value{
(data.frame) Return a \code{data.frame} corresponding to the 
metadata described.
}
\description{
Describe entities of otherEntity data objects
}
\keyword{internal}
