% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/template_physical.R
\name{template_physical}
\alias{template_physical}
\title{Describe physical attributes of data objects}
\usage{
template_physical(
  path,
  data.path = path,
  data.table = NULL,
  other.entity = NULL,
  empty = FALSE,
  write.file = TRUE,
  overwrite = FALSE
)
}
\arguments{
\item{path}{(character) Path to the metadata template directory.}

\item{data.path}{(character) Path to the data directory.}

\item{data.table}{(character; optional) Table file name. If more than one, then supply 
as a vector of character strings (e.g. 
\code{data.table = c("nitrogen.csv", "decomp.csv")}).}

\item{other.entity}{(character; optional) Name of \code{other.entity}(s) in this 
dataset. Use \code{other.entity} for all non-\code{data.table} files. 
\code{other.entity}(s) should be stored at \code{data.path}. If more 
than one, then supply as a vector of character strings (e.g. 
\code{other.entity = c('ancillary_data.zip', 'quality_control.R')}).}

\item{empty}{(logical) Whether to write an empty template file.}

\item{write.file}{(logical; optional) Whether to write the template file. Default is \code{TRUE}.}

\item{overwrite}{(logical) Overwrite the template with new content?}
}
\value{
\item{physical}{Columns:
    \itemize{
    \item{type: Object type. Can be: dataTable, otherEntity.}
    \item{entityName: A short descriptive name for the object.}
    \item{entityDescription: Description of the object.}
    \item{objectName: File name of the object.}
    \item{size: File size in bytes}
    \item{authentication: File hash value.}
    \item{authentication_method: Method of calculating the file has value. Defaults to "MD5"}
    \item{numHeaderLines: Number of header lines. Defaults to 1.}
    \item{recordDelimiter: Record delimiter (i.e. newline character).}
    \item{attributeOrientation: Can be "column" or "row". Defaults to "column".}
    \item{fieldDelimiter: Field delimiter.}
    \item{quoteCharacter: Quote character used to enclose character type data.}
    \item{formatName: Format of data object. Only for non-tabular data.}
    \item{url: The publicly accessible URL from which the data object can be downloaded.}
    \item{numberOfRecords: Number of records in the data object}
    }
}
}
\description{
This is an optional template that can be used to define the 
physical attributes of data objects. The \code{make_eml()} function will 
automatically calculate these values if the physical template is absent or 
empty.

Use this template to define physical attributes of remote data that cannot 
be accessed by \code{make_eml()}.
}
\details{
\code{make_eml()} will calculate values for empty fields and ignore
fields containing NA. Note, some fields only apply to one object type.
}
\examples{
\dontrun{

}

}
