% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/template_raster_attributes.R
\name{template_raster_attributes}
\alias{template_raster_attributes}
\title{Describe spatial raster files}
\usage{
template_raster_attributes(
  path = NULL,
  data.path = path,
  raster.files = NULL,
  empty = FALSE,
  write.file = TRUE,
  return.obj = FALSE
)
}
\arguments{
\item{path}{(character) Path to the metadata template directory.}

\item{data.path}{(character) Path to the data directory.}

\item{raster.files}{(character) File name. If more than one, then supply as a 
vector of character strings (e.g. 
\code{raster.files = c('change.tif', 'mean_votes.tif')}).}

\item{empty}{(logical) Whether to write an empty template file. Default is \code{FALSE}.}

\item{write.file}{(logical; optional) Whether to write the template file. Default is \code{TRUE}.}

\item{return.obj}{(logical; optional) Whether to return the provenance template as a data frame. Default is \code{FALSE}.}
}
\value{
\item{raster_attributes.txt}{Columns:
    \itemize{
    \item{filename: Filename of a spatial raster file}
    \item{description: Description of the raster file}
    \item{geoDescription: Description of the geographic coverage of the raster file}
    \item{url: Accessible url for download of the file}
    \item{definition: Definition of the raster attribute}
    \item{unit: Unit of the raster attribute}
    \item{numberType: ('real', 'natural', 'integer', 'whole')}
    }
}
}
\description{
Describes characteristics of spatial raster files. Use this template to facilitate the creation of the spatialRaster element in EML.
}
\examples{
\dontrun{
# Set working directory
setwd("/Users/me/Documents/data_packages/pkg_260")

# For an empty template to be filled manually
template_raster_attributes(
    path = "./metadata_templates",
    data.path = "./data_objects",
    empty = TRUE)
    
# For a template with file names
template_raster_attributes(
    path = "./metadata_templates",
    data.path = "./data_objects",
    raster.files = c('change.tif', 'mean_votes.tif'))
}

}
