% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/template_raster_catvars.R
\name{template_raster_catvars}
\alias{template_raster_catvars}
\title{Describe categorical variables associated with a raster file}
\usage{
template_raster_catvars(
  path = NULL,
  data.path = path,
  raster.file = NULL,
  write.file = TRUE
)
}
\arguments{
\item{path}{(character) Path to the metadata template directory.}

\item{data.path}{(character) Path to the data directory.}

\item{raster.file}{(character) File names of raster files. If more than one, then supply as
a vector.}

\item{write.file}{(logical; optional) Whether to write the template file.}
}
\value{
\item{catvars_*.txt}{Columns:
    \itemize{
    \item{attributeNames: Variable name}
    \item{code: Categorical variable}
    \item{definition: Definition of categorical variable}
    }
}
}
\description{
Describes the categorical variable represented in a raster file.
}
\examples{
\dontrun{
# Set working directory
setwd("/Users/me/Documents/data_packages/pkg_260")

# For files containing categorical variables as classified in the vector attributes template
template_raster_catvars(
  path = "./metadata_templates",
  data.path = "./data_objects",
  raster.file = setNames(
    c("geotiff_test_file.tif"), 
    c("attributes_geotiff_test_file.txt")
  ))
}

}
