% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/template_raster_information.R
\name{template_raster_information}
\alias{template_raster_information}
\title{Create template for information need to create EML element of raster file}
\usage{
template_raster_information(
  path = NULL,
  data.path = path,
  raster.file = NULL,
  write.file = TRUE
)
}
\arguments{
\item{path}{(character) Path to the metadata template directory.}

\item{data.path}{(character) Path to the data directory.}

\item{raster.file}{(character) File names of raster files. If more than one, then supply as
a vector.}

\item{write.file}{(logical; optional) Whether to write the template file.}
}
\value{
\item{information_*}{Columns:
    \itemize{
    \item{spatialReference: Coordinate system used.}
    \item{horizontalAccuracy: Accuracy of horizontal locational mesurements}
    \item{verticalAccuracy: Accuracy of vertical locational mesurements}
    \item{cellSizeXDirection: Width of cells in the x direction}
    \item{cellSizeYDirection: Width of cells in the y direction}
    \item{numberOfBands: Number of bands in the raster}
    \item{rasterOrigin: Location of the pixel with the minimum x and y values. 
          Valid options are : "Upper Left", "Lower Left", "Upper Right", "Lower Right"}
    \item{rows: Maximum number of objects in the y axis}
    \item{columns: Maximum number of objects in the x axis}
    \item{verticals: Maximum number of objects in the z axis}
    \item{cellGeometry: Geometric representation of the cell's content. 
          Valid options are : "pixel" (cell value representative of the entire cell) and 
          "matrix" (cell value representative of a single point within the cell).}
    }
}
}
\description{
Create a template containing all mandatory information need to create spatialRaster EML element.
}
\examples{
\dontrun{
# Set working directory
setwd("/Users/me/Documents/data_packages/pkg_260")

template_vector_information(
  path = "./metadata_templates",
  data.path = "./data_objects",
  vector.file = "geotiff_test_file.tif")
}

}
