% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/template_raster_variables.R
\name{template_raster_variables}
\alias{template_raster_variables}
\title{Describe categorical variables associated with a raster file}
\usage{
template_raster_variables(
  path = NULL,
  data.path = path,
  write.file = TRUE,
  empty = FALSE,
  return.obj = FALSE
)
}
\arguments{
\item{path}{(character) Path to the metadata template directory.}

\item{data.path}{(character) Path to the data directory.}

\item{write.file}{(logical; optional) Whether to write the template file.}

\item{empty}{(logical) Whether template should be empty. If FALSE, raster_attrbutes.txt must exist and have at least one categorical type specified.}

\item{return.obj}{(logical) Whether to return table as an R object.}
}
\value{
\item{raster_catvars.txt}{Columns:
    \itemize{
    \item{filename: Filename of a spatial raster file}
    \item{code: Categorical variable}
    \item{definition: Definition of categorical variable}
    }
}
}
\description{
Describes the categorical variable represented in a raster file. Use if any raster files have numberType of "categorical" in raster_attributes.txt.
}
\examples{
\dontrun{
# Set working directory
setwd("/Users/me/Documents/data_packages/pkg_260")

# For raster files containing categorical variables as classified in the raster attributes template
template_raster_variables(
  path = "./metadata_templates",
  data.path = "./data_objects")
}
}
