% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/template_spatial_attributes.R
\name{template_spatial_attributes}
\alias{template_spatial_attributes}
\title{Describe spatial files}
\usage{
template_spatial_attributes(
  path = NULL,
  data.path = path,
  spatial.type = "raster",
  spatial.files = NULL,
  empty = FALSE,
  write.file = TRUE,
  return.obj = FALSE
)
}
\arguments{
\item{path}{(character) Path to the metadata template directory.}

\item{data.path}{(character) Path to the data directory.}

\item{spatial.type}{(character) Type of spatial data.
Currently Supported:
    \itemize{
    \item{raster: Filename of raster data}
    \item{shape: Filename (without extension) of a shapefile}
    }}

\item{spatial.files}{(character) File name. If more than one, then supply as a 
vector of character strings (e.g. 
\code{spatial.files = c('change.tif', 'mean_votes.tif')}).}

\item{empty}{(logical) Whether to write an empty template file. Default is \code{FALSE}.}

\item{write.file}{(logical; optional) Whether to write the template file. Default is \code{TRUE}.}

\item{return.obj}{(logical; optional) Whether to return the provenance template as a data frame. Default is \code{FALSE}.}
}
\value{
\item{raster_attributes.txt}{Columns:
    \itemize{
    \item{filename: Filename of a raster}
    \item{description: Description of the spatial file}
    \item{geoDescription: Description of the geographic coverage of the spatial file}
    \item{url: Accessible url for download of the file}
    \item{definition: Definition of an attribute}
    \item{unit: Unit of an attribute}
    \item{numberType: ('real', 'natural', 'integer', 'whole', 'categorical')}
    }
}

\item{shape_attributes.txt}{Columns:
    \itemize{
    \item{extname: Filename without extension of a shapefile}
    \item{root_dir: Name of directory (within a specified path) that holds shapefile components}
    \item{description: Description of the spatial file}
    \item{geoDescription: Description of the geographic coverage of the spatial file}
    \item{url: Accessible url for download of the file}
    \item{overwrite: Allow the build_shape_element function to overwrite existing directory? (TRUE or FALSE)}
    }
}
}
\description{
Describes characteristics of spatial raster or spatial files. Use this template to facilitate the creation of the \code{<spatialRaster>} or \code{<spatialVector>} element in EML.
}
\examples{
\dontrun{
# Set working directory
setwd("/Users/me/Documents/data_packages/pkg_260")

# For an empty template to be filled manually
template_raster_attributes(
    path = "./metadata_templates",
    data.path = "./data_objects",
    empty = TRUE)
    
# For a template with file names
template_raster_attributes(
    path = "./metadata_templates",
    data.path = "./data_objects",
    raster.files = c('change.tif', 'mean_votes.tif'))
}

}
