% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/template_vector_catvars.R
\name{template_vector_catvars}
\alias{template_vector_catvars}
\title{Describe categorical variables associated with a vector file}
\usage{
template_vector_catvars(
  path = NULL,
  data.path = path,
  vector.file = NULL,
  write.file = TRUE
)
}
\arguments{
\item{path}{(character) Path to the metadata template directory.}

\item{data.path}{(character) Path to the data directory.}

\item{vector.file}{(character) File names of vector files. If more than one, then supply as
a vector.}

\item{write.file}{(logical; optional) Whether to write the template file.}
}
\value{
\item{catvars_*.txt}{Columns:
    \itemize{
    \item{attributeNames: Variable name}
    \item{code: Categorical variable}
    \item{definition: Definition of categorical variable}
    }
}
}
\description{
Describes the categorical variable represented in a vector file.
}
\examples{
\dontrun{
# Set working directory
setwd("/Users/me/Documents/data_packages/pkg_260")

# For files containing categorical variables as classified in the vector attributes template
template_vector_catvars(
  path = "./metadata_templates",
  data.path = "./data_objects",
  vector.file = setNames(
    c("shapefile_test", "geojson_test_file.GeoJSON"), 
    c("attributes_shapefile_test.txt", "attributes_geojson_test_file.txt")
  ))
}

}
