% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/template_vector_information.R
\name{template_vector_information}
\alias{template_vector_information}
\title{Create template for information need to create EML element of vector file}
\usage{
template_vector_information(
  path = NULL,
  data.path = path,
  vector.file = NULL,
  write.file = TRUE
)
}
\arguments{
\item{path}{(character) Path to the metadata template directory.}

\item{data.path}{(character) Path to the data directory.}

\item{vector.file}{(character) File names of vector files. If more than one, then supply as
a vector.}

\item{write.file}{(logical; optional) Whether to write the template file.}
}
\value{
\item{information_*}{Columns:
    \itemize{
    \item{geometry: Geometry of features in the vector file.  Accepted values are : 
          "Point", "LineString", "LinearRing", "Polygon", "MultiPoint", "MultiLineString", "MultiPolygon", "MultiGeometry"}
    \item{geometricObjectCount: Number of geometric objects}
    \item{spatialReference: Coordinate system used.}
    }
}
}
\description{
Create a template containing all mandatory information need to create spatialVector EML element.
}
\examples{
\dontrun{
# Set working directory
setwd("/Users/me/Documents/data_packages/pkg_260")

template_vector_information(
  path = "./metadata_templates",
  data.path = "./data_objects",
  vector.file = c("shapefile_test", "geojson_test_file.GeoJSON"))
}

}
