% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_templates.R
\name{validate_abstract}
\alias{validate_abstract}
\title{Validate the abstract template}
\usage{
validate_abstract(x)
}
\arguments{
\item{x}{(list) The data and metadata object returned by 
\code{template_arguments()}.}
}
\value{
\item{character}{Descriptions of any validation issues found}
    \item{NULL}{If no issues were found}
}
\description{
Validate the abstract template
}
