% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_templates.R
\name{validate_annotation}
\alias{validate_annotation}
\title{Validate the annotations template (annotations.txt)}
\usage{
validate_annotation(x)
}
\arguments{
\item{x}{(named list) The data and metadata object returned by 
\code{template_arguments()}.}
}
\value{
\item{issues}{(character) Descriptions of issues found in the template 
    grouped by "required" and "optional" criteria.}
    \item{x}{(named list) If "required" issues are found, then the template 
    is removed along with the corresponding data table.}
}
\description{
Validate the annotations template (annotations.txt)
}
\details{
Checks performed on each data tables attribute metadata:
    \itemize{
        \item{Template column names are correct}
        \item{Each annotation has a subject, predicate, and object, with
        corresponding labels and URIs}
        \item{URIs are resolvable}
    }
    
    Checks on annotation presence are only a reminder that it is possible 
    to annotate these elements.
}
