% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_templates.R
\name{validate_annotation_uri}
\alias{validate_annotation_uri}
\title{Check URIs are resolvable in the annotations template}
\usage{
validate_annotation_uri(x)
}
\arguments{
\item{x}{(list) The data and metadata object returned by 
\code{template_arguments()}.}
}
\value{
\item{character}{Description of validation issues}
    \item{NULL}{If no issues were found}
}
\description{
Check URIs are resolvable in the annotations template
}
