% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_arguments.R
\name{validate_arguments}
\alias{validate_arguments}
\title{Validate arguments of EMLassemblyline functions}
\usage{
validate_arguments(fun.name, fun.args)
}
\arguments{
\item{fun.name}{(character) Function name passed to `validate_x` with 
`as.character(match.call()[[1]])`.}

\item{fun.args}{(named list) Function arguments and values passed to `validate_x` with 
`as.list(environment())`.}
}
\description{
Validate input arguments to `EMLassemblyline` functions.
}
\details{
Validation checks are function specific.
}
