% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_templates.R
\name{validate_categorical_variable_template_presence}
\alias{validate_categorical_variable_template_presence}
\title{Check for presence of the categorical variables template}
\usage{
validate_categorical_variable_template_presence(file.name, x)
}
\arguments{
\item{file.name}{(character) The table attributes template to apply this function to. 
Full file name is required.}

\item{x}{(named list) The data and metadata object returned by 
\code{template_arguments()}.}
}
\value{
(character or NULL) A character vector of issues if any were found, 
    otherwise NULL.
}
\description{
Check for presence of the categorical variables template
}
