% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_templates.R
\name{validate_categorical_variables}
\alias{validate_categorical_variables}
\title{Validate the categorical variables template (catvars_*.txt)}
\usage{
validate_categorical_variables(x)
}
\arguments{
\item{x}{(named list) The data and metadata object returned by 
\code{template_arguments()}.}
}
\value{
\item{issues}{(character) Descriptions of issues found in the template 
    grouped by "required" and "optional" criteria.}
    \item{x}{(named list) If "required" issues are found for a data table, 
    then the corresponding categorical variables template is removed. Some
    "optional" issues reclassified the categorical variable as "character".}
}
\description{
Validate the categorical variables template (catvars_*.txt)
}
\details{
This function validates the table attributes classified as categorical
    only for data tables that are listed in \code{x}.Checks performed by 
    this function:
    \itemize{
        \item{Categorical metadata is expected for variables classified 
        "categorical"}
        \item{Categorical variable template olumn names are correct}
        \item{Categorical codes are defined}
    }
}
\keyword{internal}
