% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eml2eal.R
\name{validate_eml_content}
\alias{validate_eml_content}
\title{Check EML for supported/unsupported content}
\usage{
validate_eml_content(eml)
}
\arguments{
\item{eml}{(character) Full path to EML file}
}
\value{
Warnings/errors/messages of unsupported content found in the EML. This supplements warnings/errors/messages returned by \code{eml2*} functions, which are more context specific.
}
\description{
Check EML for supported/unsupported content
}
\keyword{internal}
