% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_arguments.R
\name{validate_geographic_coord_desc}
\alias{validate_geographic_coord_desc}
\title{Validate geographic coordinate and description pair}
\usage{
validate_geographic_coord_desc(fun.args)
}
\arguments{
\item{fun.args}{(named list) Function arguments and their values.}
}
\value{
\item{issues}{Description of issues}
    \item{fun.args}{Updated list of function arguments}
}
\description{
Validate geographic coordinate and description pair
}
\keyword{internal}
