% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_templates.R
\name{validate_geographic_coverage}
\alias{validate_geographic_coverage}
\title{Validate geographic coverage metadata}
\usage{
validate_geographic_coverage(x)
}
\arguments{
\item{x}{(named list) The data and metadata object returned by 
\code{template_arguments()}.}
}
\value{
\item{issues}{(character) Descriptions of issues found in the template 
    grouped by "required" and "optional" criteria.}
    \item{x}{(named list) If "required" issues are found, then the template 
    is removed.}
}
\description{
Validate geographic coverage metadata
}
\details{
This function compiles geographic coverage input as 
    \code{geographic.coordinates} and \code{geographic.description} to 
    \code{make_eml()}, and listed in the geographic_coverage.txt template,
    and returns all unique values in the geographic_coverage.txt template.
    
    Checks performed by this function:
    \itemize{
        \item{Column names are correct}
        \item{Geographic coverage is not missing}
        \item{Each entry is complete}
        \item{Decimal degree is expected}
        \item{The cordinate range is valid}
    }
}
\keyword{internal}
