% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_templates.R
\name{validate_geographic_coverage_column_names}
\alias{validate_geographic_coverage_column_names}
\title{Check column names of geographic coverage template}
\usage{
validate_geographic_coverage_column_names(x)
}
\arguments{
\item{x}{(list) The data and metadata object returned by 
\code{template_arguments()}.}
}
\value{
(Error or NULL) An error if any issues were found, 
    otherwise NULL.
}
\description{
Check column names of geographic coverage template
}
