% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_templates.R
\name{validate_geographic_coverage_coordinate_format}
\alias{validate_geographic_coverage_coordinate_format}
\title{Check for decimal degree format in the geographic coverage template}
\usage{
validate_geographic_coverage_coordinate_format(x)
}
\arguments{
\item{x}{(list) The data and metadata object returned by 
\code{template_arguments()}.}
}
\value{
\item{character}{Description of validation issues}
    \item{NULL}{If no issues were found}
}
\description{
Check for decimal degree format in the geographic coverage template
}
\keyword{internal}
