% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_templates.R
\name{validate_geographic_coverage_coordinate_range}
\alias{validate_geographic_coverage_coordinate_range}
\title{Check geographic coordinate ranges}
\usage{
validate_geographic_coverage_coordinate_range(x)
}
\arguments{
\item{x}{(list) The data and metadata object returned by 
\code{template_arguments()}.}
}
\value{
\item{character}{Description of validation issues}
    \item{NULL}{If no issues were found}
}
\description{
Check geographic coordinate ranges
}
\keyword{internal}
