% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_templates.R
\name{validate_geographic_coverage_presence}
\alias{validate_geographic_coverage_presence}
\title{Check geographic coverage template presence}
\usage{
validate_geographic_coverage_presence(x)
}
\arguments{
\item{x}{(named list) The data and metadata object returned by 
\code{template_arguments()}.}
}
\value{
(character or NULL) A character vector of issues if any were found, 
    otherwise NULL.
}
\description{
Check geographic coverage template presence
}
\keyword{internal}
