% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_templates.R
\name{validate_intellectual_rights}
\alias{validate_intellectual_rights}
\title{Validate the intellectual rights template}
\usage{
validate_intellectual_rights(x)
}
\arguments{
\item{x}{(list) The data and metadata object returned by 
\code{template_arguments()}.}
}
\value{
\item{character}{Description of validation issues}
    \item{NULL}{If no issues were found}
}
\description{
Validate the intellectual rights template
}
\details{
Checks performed by this function:
    \itemize{
        \item{Intellectual rights is not empty}
    }
}
\keyword{internal}
