% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_templates.R
\name{validate_keywords}
\alias{validate_keywords}
\title{Validate the keywords template}
\usage{
validate_keywords(x)
}
\arguments{
\item{x}{(list) The data and metadata object returned by 
\code{template_arguments()}.}
}
\value{
\item{character}{Description of validation issues}
    \item{NULL}{If no issues were found}
}
\description{
Validate the keywords template
}
\details{
Checks performed by this function:
    \itemize{
        \item{Keywords is not empty}
    }
}
