% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_arguments.R
\name{validate_maintenance_description}
\alias{validate_maintenance_description}
\title{Validate maintenance description}
\usage{
validate_maintenance_description(fun.args)
}
\arguments{
\item{fun.args}{(named list) Function arguments and their values.}
}
\value{
\item{issues}{Description of issues}
    \item{fun.args}{Updated list of function arguments}
}
\description{
Validate maintenance description
}
