% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_arguments.R
\name{validate_other_entity_name}
\alias{validate_other_entity_name}
\title{Validate other entity names}
\usage{
validate_other_entity_name(fun.args)
}
\arguments{
\item{fun.args}{(named list) Function arguments and their values.}
}
\value{
\item{issues}{Description of issues}
    \item{fun.args}{Updated list of function arguments}
}
\description{
Validate other entity names
}
\keyword{internal}
