% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_templates.R
\name{validate_personnel}
\alias{validate_personnel}
\title{Validate the personnel template}
\usage{
validate_personnel(x)
}
\arguments{
\item{x}{(list) The data and metadata object returned by 
\code{template_arguments()}.}
}
\value{
\item{issues}{(character) Descriptions of issues found in the template}
    \item{x}{(list) A potentially modified \code{x} if relevant issues were
    found}
}
\description{
Validate the personnel template
}
\details{
Checks performed by this function:
    \itemize{
        \item{Personnel template is not empty}
        \item{Template column names are correct}
        \item{At least one creator and contact is listed}
        \item{All personnel have roles}
        \item{Principal investigator is present}
        \item{Project information is present}
        \item{Only one publisher is listed}
    }
    
    Checks are grouped by required and optional criteria. If any required
    checks fail, then the entire template is removed from \code{x}.
}
