% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_templates.R
\name{validate_personnel_contact}
\alias{validate_personnel_contact}
\title{Check for contact in personnel template}
\usage{
validate_personnel_contact(x)
}
\arguments{
\item{x}{(list) The data and metadata object returned by 
\code{template_arguments()}.}
}
\value{
\item{character}{Description of validation issues}
    \item{NULL}{If no issues were found}
}
\description{
Check for contact in personnel template
}
\keyword{internal}
