% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_templates.R
\name{validate_personnel_presence}
\alias{validate_personnel_presence}
\title{Validate personnel template presence}
\usage{
validate_personnel_presence(x)
}
\arguments{
\item{x}{(list) The data and metadata object returned by 
\code{template_arguments()}.}
}
\value{
\item{character}{Description of validation issues}
    \item{NULL}{If no issues were found}
}
\description{
Validate personnel template presence
}
\details{
Checks performed by this function:
    \itemize{
        \item{Personnel template is not empty}
    }
}
\keyword{internal}
