% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_templates.R
\name{validate_personnel_project}
\alias{validate_personnel_project}
\title{Check for project information in the personnel template}
\usage{
validate_personnel_project(x)
}
\arguments{
\item{x}{(list) The data and metadata object returned by 
\code{template_arguments()}.}
}
\value{
\item{character}{Description of validation issues}
    \item{NULL}{If no issues were found}
}
\description{
Check for project information in the personnel template
}
\details{
Project info should be included with each PI.
}
