% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_templates.R
\name{validate_provenance}
\alias{validate_provenance}
\title{Validate the provenance template}
\usage{
validate_provenance(x)
}
\arguments{
\item{x}{(list) The data and metadata object returned by 
\code{template_arguments()}.}
}
\value{
\item{issues}{(character) Descriptions of issues found in the template}
    \item{x}{(list) A potentially modified \code{x} if relevant issues were
    found}
}
\description{
Validate the provenance template
}
\details{
This function compiles provenance input as \code{provenance} to 
    \code{make_eml()} and listed in the provenance.txt template,
    and returns all unique values in the provenance.txt template.

    Checks performed by this function:
    \itemize{
        \item{Template column names are correct}
        \item{systemID is one of the supported system identifiers}
        \item{dataPackageID and systemID pair resolves to resource metadata in \code{systemID}}
        \item{A URL is present and resolves for external resources}
        \item{An online description is recommended for external resources}
        \item{A title is present for external resources}
        \item{A persons name, or an organization name, is present for external resources}
        \item{A creator and contact (role) is listed for each external resource}
        \item{An email contact is recommended for external resources}
    }
    
    Checks are grouped by required and optional criteria. If any required
    checks fail, then the entire template is removed from \code{x}.
}
