% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_templates.R
\name{validate_provenance_column_names}
\alias{validate_provenance_column_names}
\title{Check column names of provenance template}
\usage{
validate_provenance_column_names(x)
}
\arguments{
\item{x}{(list) The data and metadata object returned by 
\code{template_arguments()}.}
}
\value{
\item{error}{If column names are invalid}
    \item{NULL}{If no issues were found}
}
\description{
Check column names of provenance template
}
