% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_templates.R
\name{validate_provenance_data_package_id}
\alias{validate_provenance_data_package_id}
\title{Check that EDI dataPackageID matches expected format (<scope>.<identifier>.<revision>)}
\usage{
validate_provenance_data_package_id(x)
}
\arguments{
\item{x}{(list) The object argument list created by 
\code{template_arguments()}.}
}
\value{
\item{character}{Description of validation issues}
    \item{NULL}{If no issues were found}
}
\description{
Check that EDI dataPackageID matches expected format (<scope>.<identifier>.<revision>)
}
\keyword{internal}
