% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_templates.R
\name{validate_provenance_external_resource_fields}
\alias{validate_provenance_external_resource_fields}
\title{Check external resource metadata is complete in the provenance template}
\usage{
validate_provenance_external_resource_fields(x)
}
\arguments{
\item{x}{(list) The data and metadata object returned by 
\code{template_arguments()}.}
}
\value{
\item{character}{Description of validation issues}
    \item{NULL}{If no issues were found}
}
\description{
Check external resource metadata is complete in the provenance template
}
