% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_templates.R
\name{validate_provenance_individual_organization_name}
\alias{validate_provenance_individual_organization_name}
\title{Check individual and organization name of external resources in the provenance template}
\usage{
validate_provenance_individual_organization_name(x)
}
\arguments{
\item{x}{(list) The data and metadata object returned by 
\code{template_arguments()}.}
}
\value{
\item{character}{Description of validation issues}
    \item{NULL}{If no issues were found}
}
\description{
Check individual and organization name of external resources in the provenance template
}
