% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_templates.R
\name{validate_provenance_system_id}
\alias{validate_provenance_system_id}
\title{Check for allowed systemID in provenance template}
\usage{
validate_provenance_system_id(x)
}
\arguments{
\item{x}{(list) The data and metadata object returned by 
\code{template_arguments()}.}
}
\value{
\item{character}{Description of validation issues}
    \item{NULL}{If no issues were found}
}
\description{
Check for allowed systemID in provenance template
}
