% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_templates.R
\name{validate_raster_attributes}
\alias{validate_raster_attributes}
\title{Validate the raster attributes template (raster_attributes.txt)}
\usage{
validate_raster_attributes(x)
}
\arguments{
\item{x}{(named list) The data and metadata object returned by 
    \code{template_arguments()}.

TODO}
}
\description{
Validate the raster attributes template (raster_attributes.txt)
}
