% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_arguments.R
\name{validate_raster_description}
\alias{validate_raster_description}
\title{Validate spatial raster description}
\usage{
validate_raster_description(fun.args)
}
\arguments{
\item{fun.args}{(named list) Function arguments and their values.}
}
\value{
\item{issues}{Description of issues}
    \item{fun.args}{Updated list of function arguments}
}
\description{
Validate spatial raster description
}
\keyword{internal}
