% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_templates.R
\name{validate_table_attribute_false_numeric}
\alias{validate_table_attribute_false_numeric}
\title{Check numeric attributes for non-numeric values}
\usage{
validate_table_attribute_false_numeric(template.name, data.name, x)
}
\arguments{
\item{template.name}{(character) The table attributes template to apply this function 
to. Full file name is required.}

\item{data.name}{(character) The data table to apply this function to. Full file name 
is required.}

\item{x}{(list) The data and metadata object returned by 
\code{template_arguments()}.}
}
\value{
\item{character or NULL}{A description of validation issues if any are
    found, otherwise NULL.}
    \item{x}{(list) A potentially modified \code{x} if relevant issues were
    found}
}
\description{
Check numeric attributes for non-numeric values
}
