% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_templates.R
\name{validate_table_attribute_missing_value_code_quantity}
\alias{validate_table_attribute_missing_value_code_quantity}
\title{Check the quantity of missing value codes}
\usage{
validate_table_attribute_missing_value_code_quantity(file.name, x)
}
\arguments{
\item{file.name}{(character) The table attributes template to apply this function 
to. Full file name is required.}

\item{x}{(list) The data and metadata object returned by 
\code{template_arguments()}.}
}
\value{
\item{character}{Description of validation issues}
    \item{NULL}{If no issues were found}
}
\description{
Check the quantity of missing value codes
}
