% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_templates.R
\name{validate_table_attribute_name_presence}
\alias{validate_table_attribute_name_presence}
\title{Check all column names of a data table are listed}
\usage{
validate_table_attribute_name_presence(template.name, data.name, x)
}
\arguments{
\item{template.name}{(character) The table attributes template to apply this function 
to. Full file name is required.}

\item{data.name}{(character) The data table to apply this function to. Full file name 
is required.}

\item{x}{(list) The data and metadata object returned by 
\code{template_arguments()}.}
}
\value{
\item{character}{Description of validation issues}
    \item{NULL}{If no issues were found}
}
\description{
Check all column names of a data table are listed
}
\keyword{internal}
