% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_templates.R
\name{validate_table_attribute_template_column_names}
\alias{validate_table_attribute_template_column_names}
\title{Check column names of table attributes template}
\usage{
validate_table_attribute_template_column_names(file.name, x)
}
\arguments{
\item{file.name}{(character) The categorical variables template to apply this function 
to. Full file name is required.}

\item{x}{(list) The data and metadata object returned by 
\code{template_arguments()}.}
}
\value{
\item{error}{If column names are invalid}
    \item{NULL}{If no issues were found}
}
\description{
Check column names of table attributes template
}
