% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_templates.R
\name{validate_table_attributes}
\alias{validate_table_attributes}
\title{Validate the table attributes template (attributes_*.txt)}
\usage{
validate_table_attributes(x)
}
\arguments{
\item{x}{(named list) The data and metadata object returned by 
\code{template_arguments()}.}
}
\value{
\item{issues}{(character) Descriptions of issues found in the template 
    grouped by "required" and "optional" criteria.}
    \item{x}{(named list) If "required" issues are found, then the template 
    is removed along with the corresponding data table.}
}
\description{
Validate the table attributes template (attributes_*.txt)
}
\details{
Checks performed on each data tables attribute metadata:
    \itemize{
        \item{Attribute metadata is present for each data table}
        \item{Template columns are correctly named}
        \item{All columns of a table are listed as attributes}
        \item{All columns of a table are listed in the same order as 
        attributes}
        \item{All attributes are defined}
        \item{All attributes are assigned a class}
        \item{Classes are of the expected type}
        \item{Numeric classes are assigned units}
        \item{All units are from the standard dictionary or defined}
        \item{Date classed attributes have a format specifier}
        \item{Date and time format specifiers are not empty}
        \item{Only one missing value code is allowed per attribute}
        \item{Missing value code is not a blank/white space}
        \item{All missing value codes are defined}
        \item{False "numeric" attributes are reclassified as "character"}
    }
}
