% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_templates.R
\name{validate_table_encoding}
\alias{validate_table_encoding}
\title{Validate character encoding of tabular files}
\usage{
validate_table_encoding(f)
}
\arguments{
\item{f}{(character) Full path to file}
}
\value{
(character or NULL) A description of validation issues if any are 
    found, otherwise NULL.
}
\description{
Validate character encoding of tabular files
}
