% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_templates.R
\name{validate_templates}
\alias{validate_templates}
\title{Validate content of `EMLassemblyline` metadata templates}
\usage{
validate_arguments(x)
}
\arguments{
\item{x}{(ADD OBJECT CLASSES) Describe objects.}
}
\description{
Validate the content of `EMLassembline` metadata templates.
}
