% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_arguments.R
\name{validate_temporal_coverage}
\alias{validate_temporal_coverage}
\title{Validate temporal coverage}
\usage{
validate_temporal_coverage(fun.args)
}
\arguments{
\item{fun.args}{(named list) Function arguments and their values.}
}
\value{
\item{issues}{Description of issues}
    \item{fun.args}{Updated list of function arguments}
}
\description{
Validate temporal coverage
}
