% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_arguments.R
\name{validate_title}
\alias{validate_title}
\title{Validate dataset title}
\usage{
validate_title(fun.args)
}
\arguments{
\item{fun.args}{(named list) Function arguments and their values.}
}
\value{
\item{character}{Description of issues}
    \item{NULL}{If no issues were found}
}
\description{
Validate dataset title
}
