% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_arguments.R
\name{validate_title_length}
\alias{validate_title_length}
\title{Check title length}
\usage{
validate_title_length(fun.args)
}
\arguments{
\item{fun.args}{(named list) Function arguments and their values.}
}
\value{
\item{character}{Description of issues}
    \item{NULL}{If no issues were found}
}
\description{
Check title length
}
\keyword{internal}
