% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_arguments.R
\name{validate_title_presence}
\alias{validate_title_presence}
\title{Check for title presence}
\usage{
validate_title_presence(fun.args)
}
\arguments{
\item{fun.args}{(named list) Function arguments and their values.}
}
\value{
\item{character}{Description of issues}
    \item{NULL}{If no issues were found}
}
\description{
Check for title presence
}
