% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_arguments.R
\name{validate_vector_url}
\alias{validate_vector_url}
\title{Validate spatial vector url}
\usage{
validate_vector_url(fun.args)
}
\arguments{
\item{fun.args}{(named list) Function arguments and their values.}
}
\value{
\item{issues}{Description of issues}
    \item{fun.args}{Updated list of function arguments}
}
\description{
Validate spatial vector url
}
\keyword{internal}
