% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_x.R
\name{validate_x}
\alias{validate_x}
\title{Validate `x` content}
\usage{
validate_x(
      fun.name,
      fun.args
    )
}
\arguments{
\item{fun.name}{(character) Function name passed to `validate_x` with 
`as.character(match.call()[[1]])`.}

\item{fun.args}{(named list) Function arguments and values passed to `validate_x` with 
`as.list(environment())`.}
}
\description{
Validate contents of `x`, a named list object facilitating 
    inputs/outputs to `EMLassemblyline` functions.
}
\details{
Validation checks are function specific.
}
