% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/view_instructions.R
\name{view_instructions}
\alias{view_instructions}
\title{Instructions for metadata templates}
\usage{
view_instructions()
}
\value{
Upon executing this function a series of instructions are printed to the 
    Console window. Proceed to the next step by pressing <enter>. Additional 
    details for each step are listed below.
}
\description{
Instructions for the metadata templates and configuration file imported 
    to the dataset working directory by \code{import_templates}.
}
\details{
Detailed instructions for each step:
    
    \emph{"Create a working directory for your dataset and import metadata
    templates."} If you haven't already, create a working directory for the 
    dataset and name it using underscore characters to span between words 
    (e.g. gleon_chloride). This directory will contain the metadata parts 
    created in the assembly line process. Run the function 
    \code{import_templates} to import metadata templates into the working 
    directory. For help running \code{import_templates} consult the 
    functions documentation by entering \code{?import_templates} into the 
    RStudio Console window.
    
    \emph{"Move your data table(s) to the working directory."} Move your 
    data tables to the dataset working directory and name them following the 
    convention 'datasetname_datatablename', where datasetname is the name 
    of your dataset (e.g. gleon_chloride), and datatablename is the name of 
    your data table (e.g. lake_characteristics). In this example, the name 
    of the data table becomes 'gleon_chloride_lake_characteristics'.
         
    \emph{"Write an abstract for your dataset ..."} Write an abstract for 
    your dataset in the file datasetname_abstract.txt. The abstract 
    should cover what, why, when, where, and how. The template is a UTF-8 
    formatted file. Use only UTF-8 symbols and text when completing.
    
    \emph{"Place additional information ..."} Place additional info in the 
    file datasetname_additional_info.txt. This is a good place 
    to share additional text based information that doesn't fall under the 
    scope of the abstract or methods (e.g. a list of research articles or 
    theses derived from this dataset). The template is a UTF-8 formatted 
    file. Use only UTF-8 symbols and text. Delete this file if you have no 
    additional information to present.
    
    \emph{"Fill out the keywords template"} List keywords that best describe
    your dataset. Consult this resource (http://vocab.lternet.edu/vocab/vocab/index.php) 
    as you populate the keywords template. Additionally, use keywords that 
    describe your lab, station, and project (e.g. GLEON, NSF, etc.). This is 
    a tab delimited file that can be edited with a spreadsheet editor. This 
    file is UTF-8 formatted. Definitions for fields in this file:
    \itemize{
        \item \strong{keyword} A keyword describing your dataset.
        \item \strong{keywordThesaurus} A keywordThesaurus (i.e. controlled
        vocuabulary) corresponding to the keyword listed in the keyword field. 
        If the keyword is not from a thesaurus or controlled vocabulary, leave
        corresponding entry in the keywordThesaurus field blank.
    }
    
    \emph{"Select an intellectual rights license for your dataset."} Select 
    an intellectual rights license. We have two recommendations: 
    'datasetname_cc_by_4.0_intellectual_rights.txt' and 
    'datasetname_cc0_1_intellectual_rights.txt'. Do not edit the text of
    these files. Delete the file you will not be using. Rename the one you 
    will be using to datasetname_intellectual_rights (e.g. 
    datasetname_intellectual_rights).
    
    \emph{"Write the methods for your dataset ..."} Explain the methods used
     to create this dataset in the file datasetname_methods.txt. Be specific, 
     include instrument descriptions, or point to a protocol online. If this 
     dataset is a synthesis of other datasets please specify dataset origins, 
     preferably their DOI or URL plus general citation information. This 
     file is a UTF-8 formatted file. Use only UTF-8 symbols and text.
    
    \emph{"Enter personnel information for your dataset ..."} Enter 
    personnel information for your dataset in the file 
    datasetname_personnel.txt. This is a tab delimited file that can be 
    edited with a spreadsheet editor. Definitions for fields in 
    this file:
    \itemize{
        \item \strong{givenName} Given name of person.
        \item \strong{surName} Surname of person.
        \item \strong{organizationName} Name of organization the person is 
        associated with.
        \item \strong{electronicMailAddress} Email address of person.
        \item \strong{userId} ORCID of person (not required). A valid entry
        for userId is the 12 digit number separated by dashes 
        (i.e. XXXX-XXXX-XXXX-XXXX). An ORCID is like a social security 
        number for scientists. [Create one here](https://orcid.org/).
        \item \strong{role} Role of person with respect to this dataset.
        Valid entries for role are:
        \itemize{
            \item \strong{creator} Dataset creator (required)
            \item \strong{pi} Principal investigator associated with this 
            dataset (required)
            \item \strong{contact} Dataset contact (required)
            \item Any other entries into the 'role' column are acceptable and will 
            be defined under the associated party element of this dataset.
        } 
        If a person serves more than one role, duplicate this persons 
        information in another row but with the new role. This file is a 
        UTF-8 formatted file. Use only UTF-8 symbols and text.
        \item \strong{projectTitle} Title of the project this dataset was 
        created under (recommended). Project titles are only listed on lines 
        where the personnel role is \emph{pi}. If auxiliary projects were 
        involved in creating this dataset then add a new line of personnel 
        below the primary project and list the auxiliary project title and 
        associated \emph{pi}.
        \item \strong{fundingAgency} Name of the entity funding the creation
        of this dataset (recommended). Only include an entry in this column for
        rows where role = \emph{pi}.
        \item \strong{fundingNumber} Number of the grant or award that 
        supported creation of this dataset (recommended). Only include an entry 
        in this column for rows where role = \emph{pi}.
    }
    
    \emph{"Add data table attributes ..."} Add data table attributes to the 
    file datasetname_datatablename_attributes.txt. This is a tab delimited 
    file that can be edited with a spreadsheet editor. Rename the file 
    according to convention (e.g. 
    gleon_chloride_lake_characteristics_attributes, where gleon_chloride is 
    the dataset name and lake_characteristics is the table name). Create an attributes 
    file for each data table. This file is a UTF-8 formatted file. Use only 
    UTF-8 symbols and text. Instructions for completing the attribute table 
    are as follows:
    \itemize{
        \item \strong{attributeName} Attribute name (column name) as 
        it appears in the data table and in the same order as in the data 
        table.
        \item \strong{attributeDefinition} Define the attribute (column 
        name). Be specific, it can be lengthy.
        \item \strong{class} Specify the attribute (column) class. This 
        is the type of value stored under the attribute. Valid options are: 
        \itemize{
            \item \strong{numeric} For numeric variables.
            \item \strong{categorical} For categorical variables.
            \item \strong{character} For variables containing text or 
            symbols that are not categorical. 
            \item \strong{Date} For date time data.
        }
        
        The list of valid options are case sensitive. If an attribute has 
        class of "numeric" or "Date", then all values of this attribute must 
        be either numeric or date time. If any character strings are present 
        in an otherwise "numeric" attribute, this attribute must be 
        classified as "character". Similarly if any values of a "Date" 
        attribute do not match the date time format string (details below), 
        then this attribute must be classified as "character".
        \item \strong{unit} If an attributes class is numeric, then you must
        provide units. If the attribute is numeric but does not have units, 
        enter "dimensionless". If the attribute class is categorical, 
        character, or Date then leave the unit field blank. If the attribute 
        is numeric and has units search the standard unit dictionary (opened 
        in the RStudio source window) for the unit of interest and enter the 
        unit "name" as it appears in the  dictionary. If the dictionary is 
        not open run these two lines of code in the console:
        
        \code{standardUnits <- get_unitList()}
        
        \code{View(standardUnits$units)}
        
        Unit names are case 
        sensitive. If you cannot find a unit in the dictionary, enter the 
        unit in the tab delimited UTF-8 formatted file 
        datasetname_custom_units.txt. Rename this file to accord with the 
        naming convention (e.g. gleon_chloride_custom_units). If you have no 
        custom units to report then delete this file. Valid custom units 
        must be convertible to SI Units (i.e. International System of 
        Units). If it cannot be converted to SI then list it in the 
        attribute defintion and enter "dimensionless" in the unit field. To 
        create a custom unit define the:
        \itemize{
            \item \strong{id} This is equivalent to the unit name. Reference 
            the standard unit dictionary formatting.
            \item \strong{unitType} The type of unit being defined. 
            Reference the dictionary for examples.
            \item \strong{parentSI} The SI equivalent of the id you have 
            entered.
            \item \strong{multiplierToSI} This is the multiplier to convert 
            from your custom unit to the SI unit equivalent.
            \item \strong{description} A description of the custom unit. 
            Reference the dictionary for examples.
        }
        \item \strong{dateTimeFormatString} Enter the date time format 
        string for each attribute of "Date" class. Remember, a class of 
        "Date" specifies the attribute as a date, time, or datetime. Enter 
        the format string in this field. If the attribute class is not 
        "Date", leave this field blank. Below are rules for constructing 
        format strings. Additional information is listed under 
        "dateTime-eml-attribute" of the current EML specification 
        (https://knb.ecoinformatics.org/#external//emlparser/docs/index.html).
        Valid date time formats are a combination of date, time, and time 
        zone strings. Below are a set of best practice recomendations that 
        we strongly encourage you to follow, but are by no means the full 
        list of currently acceptable format strings.
        \itemize{
            \item \strong{Date format strings:} YYYY-MM-DD, YYYY, YYYYMMDD, 
            YYYY-MM, YYYYMM, YYYY-DDD, YYYYDDD; where YYYY is year, MM is 
            month, DD is day of month, and DDD is day of year.
            \item \strong{Time format strings:} hh:mm:ss.sss, hhmmss.sss,
            hh:mm:ss, hhmmss, hh:mm, hhmm, hh; where hh is hour (in 24 hr
            clock), mm is minute, ss is second, and ss.sss is decimal 
            second.
            \item\strong{Time zone format strings:} Z, +hh:mm, +hhmm, +hh,
            -hh:mm, -hhmm, -hh; where Z (capitalized) is Coordinated 
            Universal Time, and + and - denote times ahead and behind UTC
            respectively.
        }
        If reporting a date without time, select one of the date format 
        strings. If reporting a date and time, select one date and one time 
        format string and combine with a single space (e.g. 
        YYYY-MM-DD hh:mm) or with a "T" (e.g. YYYY-MM-DDThh:mm). If 
        reporting a date and time, it is recommended that a time zone 
        specifier be appended without a space (e.g. YYYY-MM-DD hh:mm-hh:mm, 
        or YYYY-MM-DDThh:mm-hh:mm).
        \item \strong{missingValueCode} If a code for 'no data' is used,
        specify it here (e.g. NA, -99999).
        \item \strong{missingValueCodeExplanation} Define the missing value 
        code here.
    }
    \emph{"Fill out the file eml_configuration.R"} Provide additional 
    information about your dataset. Detailed instructions are listed as 
    comments in this file.
    
    \emph{"Make sure all files of the working directory are closed."} Some 
    functions will error out if these files are open.
    
    \emph{Define categorical variables} If you have categorical variables 
    listed in your attributes tables, you will need to supply definitions 
    for the categorical variable codes. Run the function \code{define_catvars},
    and consult the documentation of this function \code{?define_catvars} to
    define categorical variables.
    
    \emph{Extract detailed geographic coverage} If your dataset contains 
    more than one sampling location, then you will want to communicate this
    information. Run the function \code{extract_geocoverage}, and consult
    documentation of this function \code{?extract_geocoverage} to present
    this information in your metadata.
}
\note{
You may see a series of warnings reported after `make_eml` has completed. 
    These can be ignored if your EML passes validation. If validation fails,
    open the EML file in an XML editor and look for the invalid section. 
    Often a minor tweak to the EML can be made manually to bring it into 
    compliance with the EML schema.
}
\seealso{
\code{\link{import_templates}} to import metadata templates to the 
    dataset working directory.
}
