% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{vocab_lter_id}
\alias{vocab_lter_id}
\title{Get the ID for an LTER controlled vocab term}
\usage{
vocab_lter_id(x)
}
\arguments{
\item{x}{(character) A valid term in the LTER Controlled Vocabulary.}
}
\value{
(numeric) The identification number for a LTER Controlled Vocabulary 
    term.
}
\description{
Get the identification number for a valid term in the LTER Controlled 
    Vocabulary.
}
\keyword{internal}
