% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{vocab_lter_scope}
\alias{vocab_lter_scope}
\title{Get the scope of an LTER Controlled Vocabulary term}
\usage{
vocab_lter_scope(id)
}
\arguments{
\item{id}{(numeric) An identification number of a valid term in the LTER 
Controlled Vocabulary.}
}
\value{
(character) The scope description for a LTER Controlled Vocabulary 
    term. Note, not all terms have descriptions.
}
\description{
Get the scope description for a term in the LTER Controlled Vocabulary.
}
