% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_attributes.R
\name{write_attributes}
\alias{write_attributes}
\title{Write attributes from input metadata}
\usage{
write_attributes(path)
}
\arguments{
\item{path}{A path to the dataset working directory containing completed 
metadata and configuration files.}
}
\value{
A file in the dataset working directory titled 
    \emph{datasetname_datatablename_attributes.xlsx} containing attributes 
    entered in the file 
    \emph{datasetname_datatablename_attributes_draft.xlsx} along with 
    attribute number type (i.e. whole, real, etc.), and minimum and maximum 
    values.
    
    A file in the dataset working directory titled 
    \emph{datasetname_custom_units.xlsx} definitions for custom units not 
    found in the standard unit dictionary.
}
\description{
A function for writing attributes of a data entity specified 
    in the user supplied file 
    \emph{datasetname_datatablename_attributes_draft.xlsx}.
}
\details{
The user is guided through a series of steps in the process of 
    finalizing attributes for the data table(s) of the dataset working 
    directory.
}
\seealso{
\code{\link{copy_templates}} to copy metadata templates to the 
    dataset working directory.

\code{\link{run_guide}} for guidance on completing the template 
    files.

\code{\link{write_factors}} to create the factors table if the 
    attributes table contains factors.

\code{\link{extract_geocoverage}} to extract detailed geographic 
    coordinates of sampling sites.

\code{\link{create_eml}} to write the EML file.
}
