% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_factors.R
\name{write_factors}
\alias{write_factors}
\title{Write definitions for data table factors}
\usage{
write_factors(path)

    Run this function whenever your data contain factors, and any of your 
    \emph{datasetname_datatablename_attributes_draft.xlsx} or referenced 
    data tables have changed.
}
\arguments{
\item{path}{A path to the dataset working directory containing the updated 
\emph{datasetname_datatablename_attributes.xlsx}, 
\emph{eml_configuration.R} and referenced data table files.}
}
\value{
A file in the dataset working directory titled 
    \emph{datasetname_datatablename_factors.xlsx} containing factor 
    definitions, which is called upon by the \emph{create_eml} function.
}
\description{
A function for extracting and defining data table factors 
    defined in the attributes table.
}
\details{
This function overwrites any 
    \emph{datasetname_datatablename_factors.xlsx} files you have created in 
    the dataset working directory. To prevent overwriting of a specific 
    file, temporarily move the file out of the working directory.
    
    This function identifies character class attributes from the file 
    \emph{datasetname_datatablename_attributes.xlsx} and then extracts 
    unique values of these attribute to the table 
    \emph{datasetname_datatablename_factors.xlsx}, which is then opened for 
    the user to edit.
}
\seealso{
\code{\link{copy_templates}} to copy metadata templates to the 
    dataset working directory.

\code{\link{run_guide}} for guidance on completing the template 
    files.

\code{\link{write_attributes}} to create the attributes table.

\code{\link{extract_geocoverage}} to extract detailed geographic 
    coordinates of sampling sites.

\code{\link{create_eml}} to write the EML file.
}
