% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/template_spatial_coverage.R
\name{write_shp_dir}
\alias{write_shp_dir}
\title{Write shapefile in a folder}
\usage{
write_shp_dir(shp, path, name)
}
\arguments{
\item{shp}{(sf) object to write.}

\item{path}{(character) folder in which to create a shp file. Will contain a
folder named \code{name}.}

\item{name}{(character) name of the file that will be created.}
}
\description{
Write shapefile in a folder
}
