% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{write_template}
\alias{write_template}
\title{Write template to file}
\usage{
write_template(tmplt, name, path, force = FALSE)
}
\arguments{
\item{tmplt}{(data.frame) Template}

\item{name}{(character) Template file name (including extension)}

\item{path}{(character) Path to write to}

\item{force}{(logical) Overwrite existing template?}
}
\value{
(file, logical) Template and TRUE if written
}
\description{
Write template to file
}
\details{
Only works for tabular templates.
}
