% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{write_templates}
\alias{write_templates}
\title{Write a list of templates to file}
\usage{
write_templates(tmplts, nms, path, force = FALSE)
}
\arguments{
\item{tmplts}{(data.frame) List of templates to write.}

\item{nms}{(character) Template file names (including extensions) in the 
same order as \code{tmplts}.}

\item{path}{(character) Path to write to.}

\item{force}{(logical) Overwrite existing template?}
}
\value{
(file, logical) Template and TRUE if written.
}
\description{
Write a list of templates to file
}
\details{
This is a wrapper around \code{write_template()}.
}
\keyword{internal}
