% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_spatialRaster.R
\name{build_spatialRaster}
\alias{build_spatialRaster}
\title{Build spatialRaster node}
\usage{
build_spatialRaster(
  spatial.raster,
  spatial.raster.name,
  spatial.raster.description,
  spatial.raster.geographic.description,
  rasterAttributes,
  rasterFactors = NULL
)
}
\arguments{
\item{spatial.raster}{(character) Full file path}

\item{spatial.raster.name}{(character) Name of \code{spatial.raster}}

\item{spatial.raster.description}{(character) Description of \code{spatial.raster}}

\item{spatial.raster.geographic.description}{(character) Raster geographic coverage}

\item{rasterAttributes}{(data.frame) Attributes table}

\item{rasterFactors}{(data.frame) Factor codes table}
}
\value{
(emld; list) EML spatialRaster node
}
\description{
Build spatialRaster node
}
\examples{
raster.attribs <- data.frame(
  attributeName = "cover_value",
  attributeDefinition = 'Shrub cover fraction (0-1)',
  unit = 'dimensionless',
  numberType = 'real',
  missingValueCode = 0,
  missingValueCodeExplanation = 'No data')

# (NOTE: These metadata don't pertain to this .tif)
covercode <- c('0' = "no data", '1' = "non-shrub", '2' = "shrub")
raster.factor <- data.frame(
  attributeName = "cover_code",
  code = names(covercode),
  definition = unname(covercode))

build_spatialRaster(
  spatial.raster = "C:/Users/Colin/Documents/EDI/data_sets/fix_15/data_objects/shrub_cover.tif", 
  spatial.raster.description = "1-hectare JER-CDRRC shrub cover geotiff (cover fraction from 0 to 1)", 
  spatial.raster.geographic.description = "JER and CDRRC, southern New Mexico", 
  rasterAttributes = raster.attribs, 
  rasterFactors = raster.factor,
  baseURL = "https://jornada-ldc2.jrn.nmsu.edu/tif/")

}
