% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/define_catvars.R
\name{define_catvars}
\alias{define_catvars}
\title{Define categorical variables}
\usage{
define_catvars(path)

    Run this function whenever your data contain attributes of class 
    "categorical" listed in \emph{attributes_datatablename.txt} files.
}
\arguments{
\item{path}{A path to the dataset working directory containing 
\emph{attributes_datatablename.txt} and the respecive data tables.}
}
\value{
A tab delimited UTF-8 file in the dataset working directory titled 
    \emph{catvars_datatablename.txt} containing unique codes of attributes 
    of class "categorical". Open this file in a spreadsheet editor to 
    supply definitions for listed codes.
}
\description{
Identify and define categorical variables for your data tables.
}
\details{
This function will not overwrite any 
    \emph{catvars_datatablename.txt} files you have created in 
    the dataset working directory.
    
    This function identifies "categorical" class attributes from the file 
    \emph{attributes_datatablename.txt} and extracts unique 
    values of these attributes to the table for you to define.
    
    Duplicate data file names are not allowed, even if they are different 
    file types.
}
