% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/define_catvars.R
\name{define_catvars}
\alias{define_catvars}
\title{Import templates for categorical variable definitions}
\usage{
define_catvars(path, data.path = path)
}
\arguments{
\item{path}{(character) A path to your metadata templates directory containing 
\emph{attributes_datatablename.txt}.}

\item{data.path}{(character) A path to your data directory. Don't use this argument
if your data are co-located with your metadata templates.}
}
\value{
A tab delimited UTF-8 file in the dataset working directory titled 
    \emph{catvars_datatablename.txt} containing unique codes of attributes 
    of class "categorical". Open this file in a spreadsheet editor to 
    supply definitions for listed codes.
}
\description{
Data tables containing categorical variables should report 
    code-definition pairs in the EML. To do this, make sure the  Run this function(defined in 
    attributes_*.txt files) Data tables containing cIdentify and define categorical variables for your data tables. Run this 
    function whenever your data contain attributes of class "categorical" 
    listed in \emph{attributes_datatablename.txt} files.
}
\details{
This function will not overwrite any 
    \emph{catvars_datatablename.txt} files you have created in 
    the dataset working directory.
    
    This function identifies "categorical" class attributes from the file 
    \emph{attributes_datatablename.txt} and extracts unique 
    values of these attributes to the table for you to define.
    
    Duplicate data file names are not allowed, even if they are different 
    file types.
}
