% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/define_catvars.R
\name{define_catvars}
\alias{define_catvars}
\title{Import template for categorical variables}
\usage{
define_catvars(
      path,
      data.path = path,
      write.file = TRUE,
      x = NULL
    )
}
\arguments{
\item{path}{(character) Path to the directory containing table attributes 
template(s), and where the categorical variable template(s) will be
written.}

\item{data.path}{(character) Path to the directory containing the data table with 
categorical variables.}

\item{write.file}{(logical; optional) Whether to write the categorical variable template
to \code{path}.}

\item{x}{(named list; optional) Alternative input to \code{EMLassemblyline} 
functions. Use \code{template_arguments()} to create \code{x}.}
}
\value{
\itemize{
        \item{\strong{catvars_*.txt} The categorical variable 
        template. A tab delimited table.}
        \item{If using \code{x}, then the categorical variable template is 
        added to \strong{/x/templates}.}
    }
}
\description{
Import template for defining categorical variables used in data a
    table. This function automatically extracts and returns categorical 
    codes, the definitions of which are supplied by the user. 
    \href{https://ediorg.github.io/EMLassemblyline/articles/edit_metadata_templates.html}{Instructions for editing the template.}
}
\details{
\code{define_catvars()} knows which variables are \code{categorical} 
    based on their listing in the \code{class} column of the 
    attributes.txt template. 
    
    An existing categorical variables template will not be overwritten 
    by subsequent calls to \code{define_catvars()}.
}
