% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/define_catvars.R
\name{define_catvars}
\alias{define_catvars}
\title{Import template(s) for categorical variable definitions}
\usage{
define_catvars(path, data.path = path, x = NULL, write.file = TRUE)
}
\arguments{
\item{path}{(character) Path to where the template(s) will be imported.}

\item{data.path}{(character) Path to where the data files are stored.}

\item{x}{(named list) Alternative input/output to `EMLassemblyline` functions. 
Use `make_arguments()` to create `x`.}

\item{write.file}{(logical) Write `catvars` file to `path`.}
}
\value{
\itemize{
        \item{`catvars_*.txt` A tab delimited file written to `path`
        containing codes to be defined. Each template is appended with the 
        name of the data table from which the codes were extracted.}
        \item{If using `x`, then content of `catvars_*.txt` is added to `x` 
        under `/x/templates`.}
    }
}
\description{
Import template(s) for defining categorical variable codes used in data
    tables. This function automatically extracts and returns categorical 
    codes, the definitions of which must be supplied by a human. 
    Instructions for filling out the template are at
    \url{https://clnsmth.github.io/EMLassemblyline/articles/instructions.html}.
}
\details{
`define_catvars` knows which variables are `categorical` based on their 
    listing under the `class` column in the `attributes_*.txt` file(s). 
    
    Existing template(s) will not be overwritten by subsequent calls to 
    `define_catvars()`.
}
