% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detect_non_utf8.R
\name{detect_non_utf8}
\alias{detect_non_utf8}
\title{Detect non-UTF-8 encoded character strings}
\usage{
detect_non_utf8(x = "", x.type = "")
}
\arguments{
\item{x}{A variable containing character strings.}

\item{x.source}{File name from which x was created. This value is referenced when 
reporting errors to the user and prompting them to fix these errors.
Most often this should be a file name the user can modify.}
}
\description{
Identify non-UTF-8 encoded character strings and report location 
    relative to user supplied inputs.
}
