% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_geocoverage.R
\name{extract_geocoverage}
\alias{extract_geocoverage}
\title{Import template for geographic coverage}
\usage{
extract_geocoverage(
      path, 
      data.path = path, 
      data.table, 
      lat.col, 
      lon.col, 
      site.col, 
      x = NULL, 
      write.file = TRUE
    )
}
\arguments{
\item{path}{(character) Path to where the template(s) will be imported.}

\item{data.path}{(character) Path to where the data files are stored.}

\item{data.table}{(character) Data table containing geographic coordinates represented in 
decimal degrees, where latitudes south of the equator and longitudes 
west of the prime meridian are negative.}

\item{lat.col}{(character) Name of latitude column.}

\item{lon.col}{(character) Name of longitude column.}

\item{site.col}{(character) Name of site column, where site is the name of the location
specified by `lat.col` and `lon.col`.}

\item{x}{(named list) Alternative input/output to `EMLassemblyline` functions. 
Use \code{template_arguments} to create `x`.}

\item{write.file}{(logical) Write `geographic_coverage.txt` to `path`.}
}
\value{
\itemize{
        \item{`geographic_coverage.txt` A tab delimited file written to 
        `path` containing geographic coordinates and corresponding site 
        names.}
        \item{If using `x`, then content of `geographic_coverage.txt` is 
        added to `x` under `/x/templates`}
    }
}
\description{
Import template for describing the geographic coverage of a dataset.
    The content of this template is automatically extracted from a data 
    table containing geographical coordinates.
}
\details{
Existing templates will not be overwritten by subsequent calls to 
    `extract_geocoverage`.
}
\examples{
# Set working directory
setwd(tempdir())

# Create data package directory "edi_255"
file.copy(
  from = system.file('/examples/edi_255', package = 'EMLassemblyline'),
  to = '.',
  recursive = TRUE
)

# View directory contents (NOTE: geographic_coverage.txt doesn't exist)
dir('./edi_255/metadata_templates')

# Extract geographic coverage
extract_geocoverage(
  path = './edi_255/metadata_templates',
  data.path = './edi_255/data_objects',
  data.table = 'nitrogen.csv',
  site.col = 'site_name',
  lat.col = 'site_lat',
  lon.col = 'site_lon'
)

# View directory contents (NOTE: geographic_coverage.txt exists)
dir('./edi_255/metadata_templates')

# Rerunning extract_geocoverage() does not overwrite files
extract_geocoverage(
  path = './edi_255/metadata_templates',
  data.path = './edi_255/data_objects',
  data.table = 'nitrogen.csv',
  site.col = 'site_name',
  lat.col = 'site_lat',
  lon.col = 'site_lon'
)
    
}
