% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_geocoverage.R
\name{extract_geocoverage}
\alias{extract_geocoverage}
\title{Import template for geographic coverage}
\usage{
extract_geocoverage(path, data.path = path, data.file, lat.col, 
    lon.col, site.col)
}
\arguments{
\item{path}{(character) Path to where the template(s) will be imported.}

\item{data.path}{(character) Path to where the data files are stored.}

\item{data.file}{(character) File containing geographic coordinates represented in 
decimal degrees, where latitudes south of the equator and longitudes 
west of the prime meridian are negative.}

\item{lat.col}{(character) Name of latitude column.}

\item{lon.col}{(character) Name of longitude column.}

\item{site.col}{(character) Name of site column, where site is the name of the location
specified by `lat.col` and `lon.col`.}
}
\value{
`geographic_coverage.txt` A tab delimited file containing geographic
    coordinates and corresponding site names.
}
\description{
Import template for describing the geographic coverage of a dataset.
    The content of this template is automatically extracted from a data 
    table containing geographical coordinates.
}
\details{
Existing template will not be overwritten by subsequent calls to 
    `extract_geocoverage`.
}
