% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_geocoverage.R
\name{extract_geocoverage}
\alias{extract_geocoverage}
\title{Import templates for geographic coverage}
\usage{
extract_geocoverage(path, data.path = path, data.file, lat.col, 
    lon.col, site.col)
}
\arguments{
\item{path}{A path to the metadata directory.}

\item{data.path}{A path to the directory containing the data table with geographic 
information. Don't use this argument if the data table is located at 
the path argument listed above.}

\item{data.file}{Name of the input data table containing geographic coverage data.}

\item{lat.col}{Name of latitude column. Values of this column must be in decimal 
degrees. Latitudes south of the equator must be prefixed with a minus 
sign (i.e. dash, "-").}

\item{lon.col}{Name of longitude column. Values of this column must be in decimal 
degrees. Longitudes west of the prime meridian must be prefixed with a 
minus sign (i.e. dash, "-").}

\item{site.col}{Name of site column. This column lists site specific names to be 
associated with the geographic coordinates.}
}
\value{
A tab delimited file in the dataset working directory 
    titled \emph{geographic_coverage.txt} and containing decimal degree 
    latitude, decimal degree longitude, and site name.
}
\description{
Extract detailed geographic coverage (latitude, longitude, and site 
    name) to be included in the EML.
}
