% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_templates.R
\name{import_templates}
\alias{import_templates}
\title{Import templates for core metadata}
\usage{
import_templates(path, data.path = path, license, data.files)
}
\arguments{
\item{path}{(character) Path to where the templates will be imported.}

\item{data.path}{(character) Path to where the data files are stored.}

\item{license}{(character) License under which the data will be released. Use "CC0" 
(\url{https://creativecommons.org/publicdomain/zero/1.0/}), 
or "CCBY" (\url{https://creativecommons.org/licenses/by/4.0/}).}

\item{data.files}{(character) Names of tabular data files. Names of non-tabular data are
entered as arguments to `make_eml`.}
}
\value{
\itemize{
        \item{`abstract.txt`} Template for the dataset abstract.
        \item{`additional_info.txt`} Template for miscellaneous dataset
        information.
        \item{`attributes_*.txt`} Template(s) for data table attribute 
        metadata, populated with some automatically extracted metadata 
        content from the data files.
        \item{`bounding_boxes.txt`} Template for the dataset geographic 
        bounding coordinates.
        \item{`custom_units.txt`} Template for defining non-standard units 
        used in the dataset.
        \item{`intellectual_rights.txt`} The selected intellectual rights 
        license.
        \item{`keywords.txt`} Template for dataset keywords.
        \item{`methods.txt`} Template for dataset methods.
        \item{`personnel.txt`} Template for dataset personnel and funding 
        metadata.
    }
}
\description{
Import template files for storage of core metadata. Some templates are 
    populated with content detected by automated metadata extraction 
    methods. The remainder will have to be manually entered by a human. 
    Instructions for filling out the templates are at
    \url{https://clnsmth.github.io/EMLassemblyline/articles/instructions.html}.
}
\details{
Existing templates will not be overwritten by subsequent calls to 
    `import_templates`.
}
\examples{
    # Create a directory structure for an example dataset
    
    # Copy example data tables to /dataset/data
    
    # Import templates to /dataset/metadata_templates
    
}
