% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_templates.R
\name{import_templates}
\alias{import_templates}
\title{Import files for storing metadata content}
\usage{
import_templates(path, data.path = path, license, data.files)
}
\arguments{
\item{path}{(character) Path to where the templates will be copied.}

\item{data.path}{(character) Path to where the data files reside.}

\item{license}{(character) The license under which your dataset is publicly released. Use "CC0" 
(\url{https://creativecommons.org/publicdomain/zero/1.0/}), 
or "CCBY" (\url{https://creativecommons.org/licenses/by/4.0/}).}

\item{data.files}{(character) A vector of character strings specifying the names of the data files
of your dataset. It is not necessary to include the file extension. E.g 
data.files = c("data.file.1", "data.file.2", "etc.").}
}
\value{
\itemize{
        \item{abstract.txt} A file for the abstract of your dataset.
        \item{additional_info.txt} A file for information about 
        your dataset, that doesn't fit in anywhere else.
        \item{attributes_*.txt} A tab delimited table for information 
        about your data tables. NOTE: each of your data files will have a 
        corresponding attributes.txt file. `import_templates` makes a 
        series of informed guesses about the attributes of your data files and 
        writes them in these files. A human needs to verify the content.
        \item{bounding_boxes.txt} A tab delimited table for geographic bounding boxes of the
        dataset.
        \item{custom_units.txt} A tab delimited table for non-standard units used in 
        your dataset.
        \item{intellectual_rights.txt} Contains the intellectual rights license.
        \item{keywords.txt} A tab delimited table for keywords.
        \item{methods.txt} A text file for the dataset methods
        \item{personnel.txt} A tab delimited table for information about 
        personnel associated with the dataset.
    }
}
\description{
The template files contain the content that will be rendered into EML.
    They contain a mix of automatically extracted 
    content, boiler plate content, and manually entered human content.
}
\details{
New templates will not be imported if template files already exist in 
    the directory.
}
\examples{
    # Create a directory structure for an example dataset
    
    # Copy example data tables to /dataset/data
    
    # Import templates to /dataset/metadata_templates
    
}
