% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_templates.R
\name{import_templates}
\alias{import_templates}
\title{Import metadata templates}
\usage{
import_templates(path = "", license = "", data.files = c("data.file.1", "data.file.2", "etc."))
}
\arguments{
\item{path}{A character string specifying a path to the dataset working directory
(e.g. "C:/Users/Colin/Documents/data_sets/gleon_chloride").}

\item{license}{A character string specifying the license for your dataset. Select "CC0" 
or "CCBY". Additional information about these licenses are listed below 
under "details".}

\item{data.files}{A list of character strings specifying the names of the data files
of your dataset. It is not necessary to include the file extension.}
}
\value{
\strong{abstract.txt} A text file for the abstract of your dataset. Edit 
    this file in a text editor. Do not include special characters, symbols, 
    or formatting. Keep it simple! Describe your methods in plain text. 
    Remove any smart quotes or other symbols specific to Microsoft Office.
    
    \strong{additional_info.txt} A text file for additional information about 
    your dataset.
    
    \strong{attributes_datafilename.txt} A tab delimited table for information 
    about your data tables. Edit this file in a spread sheet editor, DO NOT 
    edit in a text editor. NOTE: each of your data files will have a 
    corresponding attributes.txt file. \code{import_templates} makes a 
    series of informed guesses about the attributes of your data files and 
    writes them in these files. You must verify this information is correct.
    
    \strong{configuration.R} A file for supplying additional parameters to 
    functions used in the assembly line. Edit this file in RStudio.
    
    \strong{custom_units.txt} A tab delimited table for custom units used in 
    your data that are not defined in the standard unit dictionary. Edit 
    this file in a spread sheet editor, DO NOT edit in a text editor.
    
    \strong{instructions.pdf} Step-by-step instructions on how to operate the 
    assembly line.
    
    \strong{intellectual_rights.txt} The selected intellectual rights license 
    for your dataset. DO NOT edit the text of this file.
    
    \strong{keywords.txt} A tab delimited table for keywords. Edit this file 
    in a spread sheet editor, DO NOT edit in a text editor.
    
    \strong{methods.txt} A text file for the methods followed in 
    collecting/creating your dataset. Edit this file in a text editor. Do 
    not include special characters, symbols, or formatting. Keep it simple! 
    Describe your methods in plain text. Remove any smart quotes or other 
    symbols specific to Microsoft Office.
    
    \strong{my_workflow.R} A blank R script for you to build an assembly 
    line workflow, which can be revisited or modified for future assembly 
    line runs.
    
    \strong{personnel.txt} A tab delimited table for information about 
    personnel associated with your dataset. Edit this file in a spread sheet 
    editor, DO NOT edit in a text editor.
}
\description{
Run this function to import metadata templates for your dataset. 
    Information entered in these templates will be rendered into EML by 
    \code{make_eml} during the last step of the assembly line process.
}
\details{
New templates will not be imported if template files already exist in 
    the directory.
    
    Here is the text of the recommended intellectual rights licenses.
    \itemize{
        \item \strong{CCO} This data package is released to the “public 
        domain” under Creative Commons CC0 1.0 “No Rights Reserved” (see: 
        https://creativecommons.org/publicdomain/zero/1.0/). It is 
        considered professional etiquette to provide attribution of the 
        original work if this data package is shared in whole or by 
        individual components. A generic citation is provided for this data 
        package on the website https://portal.edirepository.org (herein 
        “website”) in the summary metadata page. Communication (and 
        collaboration) with the creators of this data package is recommended 
        to prevent duplicate research or publication. This data package (and 
        its components) is made available “as is” and with no warranty of 
        accuracy or fitness for use. The creators of this data package and 
        the website shall not be liable for any damages resulting from 
        misinterpretation or misuse of the data package or its components. 
        Periodic updates of this data package may be available from the 
        website. Thank you.
        \item \strong{CCBY} This information is released under the Creative 
        Commons license - Attribution - CC BY 
        (https://creativecommons.org/licenses/by/4.0/). The consumer of 
        these data ("Data User" herein) is required to cite it appropriately 
        in any publication that results from its use. The Data User should 
        realize that these data may be actively used by others for ongoing 
        research and that coordination may be necessary to prevent duplicate 
        publication. The Data User is urged to contact the authors of these 
        data if any questions about methodology or results occur. Where 
        appropriate, the Data User is encouraged to consider collaboration 
        or co-authorship with the authors. The Data User should realize that 
        misinterpretation of data may occur if used out of context of the 
        original study. While substantial efforts are made to ensure the 
        accuracy of data and associated documentation, complete accuracy of 
        data sets cannot be guaranteed. All data are made available "as is." 
        The Data User should be aware, however, that data are updated 
        periodically and it is the responsibility of the Data User to check 
        for new versions of the data. The data authors and the repository 
        where these data were obtained shall not be liable for damages 
        resulting from any use or misinterpretation of the data. Thank you.
    }
}
