% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_templates.R
\name{issues}
\alias{issues}
\title{View validation issues}
\usage{
issues()
}
\value{
A message listing any validation issues
}
\description{
Validation functions return a list of validation issues to the global 
    environment in an \code{issues} object. The \code{view_issues()} 
    function wraps these issues in \code{message()} to provide a human 
    readable form.
}
