% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_eml.R
\name{make_eml}
\alias{make_eml}
\title{Make EML}
\usage{
make_eml(path)
}
\arguments{
\item{path}{A path to the dataset working directory containing the completed 
metadata templates, \emph{configuration.R}, 
\emph{catvars_datatablename.txt} (if categorical variables 
are present), and \emph{geographic_coverage.txt} (if reporting detailed 
geographic coverage).}

\item{user.id}{A character string specifying your EDI data repository user ID. If you
don't have one, contact EDI (info@environmentaldatainitiative.org) to 
get one, or don't use this argument when running \code{make_eml}.}

\item{author.system}{A character string specifying the author system your user ID is 
associated with. If you are}
}
\value{
Validation results printed to the RStudio \emph{Console}.
    
    An EML file written to the dataset working directory titled 
    \emph{packageID.xml}.
}
\description{
Translate user supplied metadata templates into the EML, validate the 
    EML, and write to file.
}
\details{
If validation fails, open the EML document in a .xml editor to identify 
    the source of error. Often the error is small and quickly resolved with 
    the aid of an editors schema validator.
}
