% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{read_data_objects}
\alias{read_data_objects}
\title{Read data objects}
\usage{
read_data_objects(data.path)
}
\arguments{
\item{path}{(character) Path to a directory containing data objects, and
potentially metadata template files.}
}
\value{
(list) Named list of parsed data objects, or NA when the object
could not be read.
}
\description{
Attempts to read data objects through a series of parsers informed by the
data object mime type.
}
\examples{
\dontrun{

}

}
\keyword{internal}
