% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_guide.R
\name{run_guide}
\alias{run_guide}
\title{User guide}
\usage{
run_guide()
}
\value{
Prompts the user to complete a series of tasks before running other functions
of the EMLtools package.
}
\description{
A guide for creating files required by functions of the EMLtools package.
}
\seealso{
\code{\link{write_attributes}} for writing data attributes

\code{\link{write_factors}} for writing data factors

\code{\link{create_eml}} for creating EML

\code{\link{copy_templates}} for creating template files
}
