% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/template_annotations.R
\name{template_annotations}
\alias{template_annotations}
\title{Create the annotations template}
\usage{
template_annotations(
      path,
      data.path = path,
      data.table = NULL,
      other.entity = NULL
    )
}
\arguments{
\item{path}{(character) Path to the metadata template directory and where 
annotations.txt will be written.}

\item{data.path}{(character; optional) Path to the data directory. Defaults to 
\code{path}.}

\item{data.table}{(character; optional) Table name. If more than one, then supply as a 
vector of character strings (e.g. 
\code{data.table = c('nitrogen.csv', 'decomp.csv')}).}

\item{other.entity}{(character; optional) Other entity name. If more than one, then supply 
as a vector of character strings (e.g. 
\code{other.entity = c('maps.zip', 'analysis_script.R')}).}
}
\value{
\strong{annotations.txt} The tab delimited annotations template.
}
\description{
Annotate your metadata with terms from an ontology. Run this function 
    after all your metadata templates are complete
}
\details{
This function extracts annotatable metadata from the template files
    and assigns default annotations for the predicate labels and URIs. The 
    user manually assigns the object labels and URIs.
    
    \code{data.table} and \code{other.entity} are added to annotations.txt 
    and assigned UUIDs.
    
    Unique persons within personnel.txt are added to annotations.txt and 
    assigned UUIDs, which are also added to the \code{id} field of 
    personnel.txt.
}
\examples{
# Add completed metadata templates to a temporary directory
file.copy(
 from = system.file("/examples/pkg_260/metadata_templates", package = "EMLassemblyline"),
 to = tempdir(),
 recursive = TRUE
)
unlink("./metadata_templates/annotations.txt", force = TRUE)
setwd(paste0(tempdir(), "/metadata_templates"))

# Create the annotations template
template_annotations(
  path = "."
)

# View directory contents
dir(".")

# View template contents
df <- data.table::fread("./annotations.txt")
df

# Remove the temporary directory
setwd(tempdir())
unlink("./metadata_templates", recursive = TRUE, force = TRUE)
    
}
