% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/template_arguments.R
\name{template_arguments}
\alias{template_arguments}
\title{Initialize inputs to \code{EMLassemblyline} functions}
\usage{
template_arguments(path = NULL, data.path = NULL, data.table = NULL,
  other.entity = NULL, sep = NULL, empty = FALSE)
}
\arguments{
\item{path}{(character) Path to the directory containing \code{EMLassemblyline} 
metadata templates.}

\item{data.path}{(character) Path to the directory containing \code{data.table} and 
\code{other.entity} data objects.}

\item{data.table}{(character) File names of data tables. If more than one, then supply as 
a vector (e.g. \code{data.table = c('decomp.csv', 'nitrogen.csv')}).}

\item{other.entity}{(character) File names of other entity data objects. If more than one, 
then supply as a vector (e.g. 
\code{other.entity = c('ancillary_data.zip', 'processing_and_analysis.R')}).}

\item{sep}{(character) Data table field delimiter. Use this if this function fails
to read your \code{data.table}.}

\item{empty}{(logical) Initialize the output with a set of empty metadata templates?
This option is useful when wanting to transfer metadata directly into
the template files without having to call the templating functions.}
}
\value{
(named list) A list of all \code{EMLassemblyline} arguments, specified 
    metadata templates and data objects.
}
\description{
Initialize the full list of possible inputs to \code{EMLassemblyline}
    functions (i.e. function arguments, metadata templates, and data 
    objects). This enables bridging of upstream metadata and data sources 
    (e.g. metabases, databases, Excel files, web forms, etc.).
}
\examples{
# Output arguments and empty versions of metadata templates

r <- template_arguments(empty = TRUE)

# Output arguments, completed versions of metadata templates, and data objects

mydir <- file.copy(
  system.file('/examples/pkg_260', package = 'EMLassemblyline'), 
  tempdir(),
  recursive = TRUE)

setwd(paste0(tempdir(), "/pkg_260"))

r <- template_arguments(
  path = "./metadata_templates",
  data.path = "./data_objects",
  data.table = c("decomp.csv", "nitrogen.csv"),
  other.entity = c("ancillary_data.zip", "processing_and_analysis.R"))
    
}
