% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/template_categorical_variables.R
\name{template_categorical_variables}
\alias{template_categorical_variables}
\title{Create categorical variables template}
\usage{
template_categorical_variables(
      path,
      data.path = path,
      write.file = TRUE,
      x = NULL
    )
}
\arguments{
\item{path}{(character) Path to the directory containing table attributes 
template(s), and where the categorical variable template(s) will be
written.}

\item{data.path}{(character) Path to the directory containing the data table with 
categorical variables.}

\item{x}{(named list; optional) Alternative input to \code{EMLassemblyline} 
functions. Use \code{template_arguments()} to create \code{x}.}

\item{write.file}{(logical; optional) Whether to write the categorical variable template
to \code{path}.}
}
\value{
\itemize{
        \item{\strong{catvars_*.txt} The categorical variable 
        template. A tab delimited table.}
        \item{If using \code{x}, then the categorical variable template is 
        added to \strong{/x/templates}.}
    }
}
\description{
Create template for defining categorical variable codes used in data
    tables. This function automatically extracts and returns categorical 
    codes, the definitions of which must be supplied by the user. 
    \href{https://ediorg.github.io/EMLassemblyline/articles/edit_metadata_templates.html}{Instructions for editing the template.}
}
\details{
\code{define_catvars()} knows which variables are \code{categorical} 
    based on their listing in the \code{class} column of the 
    attributes.txt template. 
    
    An existing categorical variables template will not be overwritten 
    by subsequent calls to \code{define_catvars()}.
}
\examples{
# Set working directory
setwd(tempdir())

# Create data package directory "edi_255"
file.copy(
  from = system.file('/examples/edi_255', package = 'EMLassemblyline'),
  to = '.',
  recursive = TRUE
)

# View directory contents (NOTE: catvars_*.txt files don't exist)
dir('./edi_255/metadata_templates')

# Template categorical variables
template_categorical_variables(
  path = './edi_255/metadata_templates',
  data.path = './edi_255/data_objects'
)

# View directory contents (NOTE: catvars_*.txt files exist)
dir('./edi_255/metadata_templates')

# Rerunning template_categorical_variables() does not overwrite files
template_categorical_variables(
  path = './edi_255/metadata_templates',
  data.path = './edi_255/data_objects'
)

}
