% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/template_categorical_variables.R
\name{template_categorical_variables}
\alias{template_categorical_variables}
\title{Create categorical variables template}
\usage{
template_categorical_variables(
      path,
      data.path = path,
      x = NULL,
      write.file = TRUE
    )
}
\arguments{
\item{path}{(character) Path to where the template(s) will be imported.}

\item{data.path}{(character) Path to where the data files are stored.}

\item{x}{(named list) Alternative input/output to `EMLassemblyline` functions. 
Use \code{template_arguments} to create `x`.}

\item{write.file}{(logical) Write `catvars` file to `path`.}
}
\value{
\itemize{
        \item{`catvars_*.txt` A tab delimited file written to `path`
        containing codes to be defined. Each template is appended with the 
        name of the data table from which the codes were extracted.}
        \item{If using `x`, then content of `catvars_*.txt` is added to `x` 
        under `/x/templates`.}
    }
}
\description{
Import template(s) for defining categorical variable codes used in data
    tables. This function automatically extracts and returns categorical 
    codes, the definitions of which must be supplied by a human. 
    Instructions for filling out the template are at
    \url{https://clnsmth.github.io/EMLassemblyline/articles/instructions.html}.
}
\details{
`template_categorical_variables` knows which variables are `categorical` based on their 
    listing under the `class` column in the `attributes_*.txt` file(s). 
    
    Existing template(s) will not be overwritten by subsequent calls to 
    `template_categorical_variables()`.
}
\examples{
# Set working directory
setwd(tempdir())

# Create data package directory "edi_255"
file.copy(
  from = system.file('/examples/edi_255', package = 'EMLassemblyline'),
  to = '.',
  recursive = TRUE
)

# View directory contents (NOTE: catvars_*.txt files don't exist)
dir('./edi_255/metadata_templates')

# Template categorical variables
template_categorical_variables(
  path = './edi_255/metadata_templates',
  data.path = './edi_255/data_objects'
)

# View directory contents (NOTE: catvars_*.txt files exist)
dir('./edi_255/metadata_templates')

# Rerunning template_categorical_variables() does not overwrite files
template_categorical_variables(
  path = './edi_255/metadata_templates',
  data.path = './edi_255/data_objects'
)

}
