% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/template_core_metadata.R
\name{template_core_metadata}
\alias{template_core_metadata}
\title{Describe core features of a data package}
\usage{
template_core_metadata(
  path,
  license,
  file.type = ".txt",
  write.file = TRUE,
  x = NULL
)
}
\arguments{
\item{path}{(character) Path to the metadata template directory.}

\item{license}{(character) License to publicly release the data package under. Use
\href{https://creativecommons.org/publicdomain/zero/1.0/}{"CC0"} or 
\href{https://creativecommons.org/licenses/by/4.0/}{"CCBY"}.}

\item{file.type}{(character; optional) File type for abstract, methods, and additional info. Can be: ".txt" (plain text), ".docx" (MS Word), or ".md" (markdown). The relative benefits: ".docx" supports basic formatting (super/sub scripts, italics, symbols, accented characters) but doesn't support bulleted lists and formatted equations. ".md" supports less formatting than ".docx" but is open source. ".txt" doesn't support any formatting but is a common file type.}

\item{write.file}{(logical; optional) Whether to write the template files.}

\item{x}{(named list; optional) Alternative input to 
\code{template_core_metadatta()}. Use \code{template_arguments()} 
to create \code{x}.}
}
\value{
\item{abstract}{Describes the salient features of a dataset in a concise summary much like an abstract does in a journal article. It should cover what the data are and why they were created.}
\item{methods}{Describes the data creation methods. Includes enough detail for future users to correctly use the data. Lists instrument descriptions, protocols, etc.}
\item{keywords}{Describes the data in a small set of terms. Keywords facilitate search and discovery on scientific terms, as well as names of research groups, field stations, and other organizations. Using a controlled vocabulary or thesaurus vastly improves discovery. We recommend using the \href{http://vocab.lternet.edu/vocab/vocab/index.php}{LTER Controlled Vocabulary} when possible. Columns:
    \itemize{
    \item{keyword: One keyword per line}
    \item{keywordThesaurus: URI of the vocabulary from which the keyword originates.}
    }
}
\item{personnel}{Describes the personnel and funding sources involved in the creation of the data. This facilitates attribution and reporting. Columns:
    \itemize{
    \item{givenName: First name}
    \item{middleInitial: Middle initial}
    \item{surName: Last name}
    \item{organizationName: Organization the person belongs to}
    \item{electronicMailAddress: Email address}
    \item{userId: Persons research identifier (e.g. \href{https://orcid.org/}{ORCID}). Links a persons research profile to a data publication.}
    \item{role: Role of the person with respect to the data. Persons serving more than one role are listed on separate lines (e.g. replicate the persons info on separate lines but change the role. Valid options: "creator" (Author of the data. Will appear in the data citation.), "PI" (Principal investigator the data were created under. Will appear with project level metadata.), "contact" (A point of contact for questions about the data. Can be an organization or position (e.g. data manager). To do this, enter the organization or position name under givenName and leave middleInitial and surName empty.), Other roles (e.g. "Field Technician") will be listed as associated parties to the data.}
    \item{projectTitle: Title of project the data were created under. If ancillary projects were involved, then add as new lines below the primary project with the PIs info replicated.}
    \item{fundingAgency: Agency the project was funded by.}
    \item{fundingNumber: Grant or award number.}
    }
}
\item{intellectual_rights}{Describes how the data may be used. Releasing without restriction (\href{https://creativecommons.org/publicdomain/zero/1.0/}{CC0}) or with minimal attribution (\href{https://creativecommons.org/licenses/by/4.0/}{CC BY}) maximizes value and future use.}
\item{additional_info}{Ancillary info not captured by any of the other templates.}
}
\description{
Core information to all data packages (abstract, methods, keywords, personnel, license). Describes what the data are, why and how they were created, who was involved in their creation, and under what license the data are released.
}
\examples{
\dontrun{
# Set working directory
setwd("/Users/me/Documents/data_packages/pkg_260")

# For data licensed under CC0 and using .docx files for the abstract and methods
template_core_metadata(
  path = "./metadata_templates",
  license = "CC0",
  file.type = ".docx")
}
    
}
