% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/template_core_metadata.R
\name{template_core_metadata}
\alias{template_core_metadata}
\title{Create core metadata templates}
\usage{
template_core_metadata(
      path,
      license,
      x = NULL,
      write.file = TRUE
    )
}
\arguments{
\item{path}{(character) Path to where the templates will be imported.}

\item{license}{(character) License under which the data will be released. Use "CC0" 
(\url{https://creativecommons.org/publicdomain/zero/1.0/}), 
or "CCBY" (\url{https://creativecommons.org/licenses/by/4.0/}).}

\item{x}{(named list) Alternative input/output to `EMLassemblyline` functions. 
Use \code{template_arguments} to create `x`.}

\item{write.file}{(logical) Write template files to `path`.}
}
\value{
\itemize{
        \item{`abstract.txt`} Template for the dataset abstract.
        \item{`additional_info.txt`} Template for miscellaneous dataset
        information.
        \item{`intellectual_rights.txt`} The selected intellectual rights 
        license.
        \item{`keywords.txt`} Template for dataset keywords.
        \item{`methods.txt`} Template for dataset methods.
        \item{`personnel.txt`} Template for dataset personnel and funding 
        metadata.
        \item{If using `x`, then content of each above listed template file 
        is added to `x` under `/x/templates/`}
    }
}
\description{
Import template files for storage of core metadata. Some templates are 
    populated with content detected by automated metadata extraction 
    methods. The remainder will have to be manually entered by a human. 
    Instructions for filling out the templates are at
    \url{https://clnsmth.github.io/EMLassemblyline/articles/instructions.html}.
}
\details{
Existing templates will not be overwritten by `template_core_metadata`.
}
