% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/template_core_metadata.R
\name{template_core_metadata}
\alias{template_core_metadata}
\title{Describe core features of a data package}
\usage{
template_core_metadata(
  path,
  license,
  file.type = ".txt",
  write.file = TRUE,
  x = NULL
)
}
\arguments{
\item{path}{(character) Path to the metadata template directory.}

\item{license}{(character) License to publicly release the data package under. Use
\href{https://creativecommons.org/publicdomain/zero/1.0/}{"CC0"} or 
\href{https://creativecommons.org/licenses/by/4.0/}{"CCBY"}.}

\item{file.type}{(character; optional) File type for abstract, methods, and additional 
info metadata templates. Can be: '.txt' (plain text), '.docx' (MS Word),
or '.md' (markdown).}

\item{write.file}{(logical; optional) Whether to write the template files.}

\item{x}{(named list; optional) Alternative input to 
\code{template_core_metadatta()}. Use \code{template_arguments()} 
to create \code{x}.}
}
\value{
\itemize{
        \item{\strong{abstract} The abstract template.}
        \item{\strong{additional_info} The template for miscellaneous
        information.}
        \item{\strong{intellectual_rights.txt} The intellectual rights 
        license with the text of CC0 or CCBY.}
        \item{\strong{keywords.txt} The tab delimited keywords template.}
        \item{\strong{methods} The methods template.}
        \item{\strong{personnel.txt} The tab delimited personnel template 
        for information on persons and funding involved in the creation of 
        the data package.}
        \item{These files is written to \code{path} unless using \code{x},
        in which case the templates are added to 
        \strong{/x/templates}.}
    }
}
\description{
Core information to all data packages (abstract, methods, keywords, personnel, license). Describes what the data are, why and how they were created, who was involved in their creation, and under what license the data are released.
}
\details{
Existing templates will not be overwritten by \code{template_core_metadata()}.
}
\examples{
\dontrun{
# Set working directory
setwd("/Users/me/Documents/data_packages/pkg_260")

# For data licensed under CC0 and using .docx files for the abstract and methods
template_core_metadata(
  path = "./metadata_templates",
  license = "CC0",
  file.type = ".docx")
}
    
}
