% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/template_directories.R
\name{template_directories}
\alias{template_directories}
\title{Create data package directory}
\usage{
template_directories(
      path, 
      dir.name
    )
}
\arguments{
\item{path}{(character) Path to where the data package directory will be created.}

\item{dir.name}{(character) Name of the directory.}
}
\value{
A directory with the following structure and contents:
    \itemize{
        \item{\strong{name} Name supplied via \code{dir.name}}
        \itemize{
            \item{\strong{data_objects} Directory for data and other 
            objects to be packaged.}
            \item{\strong{metadata_templates} Directory for 
            \code{EMLassemblyline} template files.}
            \item{\strong{eml} Directory for EML files created by 
            \code{EMLassemblyline}.}
            \item{\strong{run_EMLassemblyline.R} An empty R file for 
            scripting an \code{EMLassemblyline} workflow.}
        }
    }
}
\description{
Create a directory structure for data package contents and 
    \code{EMLassemblyline} files. Any organization scheme is supported. This
    is one recommendation.
}
\details{
Existing directories named with \code{dir.name} at \code{path} will not 
    be overwritten.
}
\examples{
# Set working directory
setwd(tempdir())

# Template data package directory "edi_301"
template_directories(
  path = '.',
  dir.name =  'edi_301'
)

# View directory contents
dir('./edi_301')

}
