% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/template_geographic_coverage.R
\name{template_geographic_coverage}
\alias{template_geographic_coverage}
\title{Describe where the data were collected}
\usage{
template_geographic_coverage(
  path,
  data.path = path,
  data.table,
  lat.col,
  lon.col,
  site.col,
  empty = FALSE,
  write.file = TRUE,
  x = NULL
)
}
\arguments{
\item{path}{(character) Path to the metadata template directory.}

\item{data.path}{(character) Path to the data directory.}

\item{data.table}{(character) Table name containing geographic coordinates.}

\item{lat.col}{(character) Column name containing latitude coordinates in decimal 
degrees, where latitudes south of the equator are negative.}

\item{lon.col}{(character) Column name containing longitude coordinates in decimal 
degrees, where longitudes west of the prime meridian are negative.}

\item{site.col}{(character) Column name containing site names.}

\item{empty}{(logical) Whether to write an empty template file.}

\item{write.file}{(logical; optional) Whether to write the template file.}

\item{x}{(named list; optional) Alternative input to 
\code{template_geographical_coverage()}. Use \code{template_arguments()} 
to create \code{x}.}
}
\value{
\strong{geographic_coverage.txt} The tab delimited geographic coverage 
    template. This file is written to \code{path} unless using \code{x},
    in which case the template is added to 
    \strong{/x/templates/catvars_*.txt}.
}
\description{
Use this function to extract the unique location coordinates and names from a table or to create a blank template to manually fill.
}
\details{
An existing geographic coverage template will not be overwritten by 
    subsequent calls to \code{template_geographic_coverage()}.
    
    Character encoding of metadata extracted directly from the tables are 
    converted to UTF-8 via \code{enc2utf8()}.
}
\examples{
\dontrun{
# Set working directory
setwd("/Users/me/Documents/data_packages/pkg_260")

# For a table containing site coordinates and names
template_geographic_coverage(
  path = "./metadata_templates",
  data.path = "./data_objects",
  data.table = "nitrogen.csv",
  site.col = "site_name",
  lat.col = "site_lat",
  lon.col = "site_lon")
  
# For returning an empty template to be filled manually
template_geographic_coverage(
  path = "./metadata_templates",
  empty = TRUE)
}

}
