% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/template_provenance.R
\name{template_provenance}
\alias{template_provenance}
\title{Create provenance template}
\usage{
template_provenance(
  path,
  data.path = path,
  data.table,
  lat.col,
  lon.col,
  site.col,
  empty = FALSE,
  write.file = TRUE,
  x = NULL
)
}
\arguments{
\item{path}{(character) Path to the metadata template directory.}

\item{empty}{(logical) Whether to write an empty template file.}

\item{write.file}{(logical; optional) Whether to write the template file.}

\item{package.id}{(character) One or more source data package IDs.}

\item{system.id}{(character) One or more source system IDs (data repositories) for each \code{package.id}. Currently supported repositories: Environmental Data Initiative.}
}
\value{
\item{provenance.txt}{The tab delimited provenance template written to \code{path} with the columns:
\itemize{
    \item{title - The data resource title}
    \item{givenName - A creator or contacts given name}
    \item{middleInitial - A creator or contacts middle initial}
    \item{surName - A creator or contacts middle initial}
    \item{role - "creator" or "contact" of the data resource}
    \item{organizationName - Name of organization the creator or contact belongs to}
    \item{email - Email of the creator or contact}
    \item{onlineDescription - Description of the data resource}
    \item{url - URL linking to the data resource}

}
}
}
\description{
Use this function to create provenance metadata for the input data 
    sources to this data package.
}
\details{
An existing provenance template will not be overwritten by a call to \code{template_provenance()}.
}
\examples{
# Set working directory
setwd(tempdir())

# The most common use case is import of a blank template and manual 
#completion
template_provenance(path = ".")

# Clean up
unlink('.', recursive = TRUE)

}
