% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/template_raster_attributes.R
\name{template_raster_attributes}
\alias{template_raster_attributes}
\title{Describe raster files}
\usage{
template_raster_attributes(
  path = NULL,
  data.path = path,
  raster.file = NULL,
  write.file = TRUE
)
}
\arguments{
\item{path}{(character) Path to the metadata template directory.}

\item{data.path}{(character) Path to the data directory.}

\item{raster.file}{(character) File name. If more than one, then supply as a
vector of character strings 
(e.g. \code{raster.file = c('change.tif', 'mean_votes.tif')}).}

\item{write.file}{(logical; optional) Whether to write the template file. Default is \code{TRUE}.}
}
\value{
\item{attributes_*}{Columns:
    \itemize{
    \item{attributeName: Column name}
    \item{attributeDefinition: Column definition}
    \item{class: Column class. Valid options are: "numeric" (Numeric variable), "categorical" (Categorical variable, i.e. nominal), "character" (Free text character strings, e.g. notes), "Date" (Date and time variable)}
    \item{unit: Column unit. Required for numeric classes. Select from EML's standard unit dictionary, accessible with \code{view_unit_dictionary()}. Use values in the "id" column. If not found, then define as a custom unit (see custom_units.txt).}
    \item{dateTimeFormatString: Format string. Required for Date classes. Valid format string components are: "Y" (Year), "M" (Month), "D" (Day), "h" (Hour), "m" (Minute), "s" (Second), Common separators of format string components (e.g. "-" "/" "\" ":"") are supported.}
    \item{missingValueCode: Missing value code. Required for columns containing a missing value code).}
    \item{missingValueCodeExplanation: Definition of missing value code.}
    }
}
\item{custom_units}{Describes non-standard units used in a vector attribute template. Columns:
    \itemize{
    \item{id: Unit name listed in the unit column of the vector attributes template (e.g. feetPerSecond)}
    \item{unitType: Unit type (e.g. velocity)}
    \item{parentSI: SI equivalent (e.g. metersPerSecond)}
    \item{multiplierToSI: Multiplier to SI equivalent (e.g. 0.3048)}
    \item{Abbreviation: Abbreviation (e.g. ft/s)}
    }
}
}
\description{
Describes variables and dimensions of a raster file
 (classes, units, datetime formats, missing value codes).
 Only GeoTiff files are currently supported.
}
\details{
Character encoding of metadata extracted directly from the tables are
    converted to UTF-8 via \code{enc2utf8()}.
}
\examples{
\dontrun{
# Set working directory
setwd("/Users/me/Documents/data_packages/pkg_260")

template_raster_attributes(
  path = './metadata_templates',
  data.path = './data_objects',
  raster.file = 'geotiff_test_file.tif')
}
    
}
