% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/template_table_attributes.R
\name{template_table_attributes}
\alias{template_table_attributes}
\title{Create table attributes template}
\usage{
template_table_attributes(
      path,
      data.path = path,
      data.table = NULL,
      x = NULL,
      write.file = TRUE
    )
}
\arguments{
\item{path}{(character) Path to where the templates will be imported.}

\item{data.path}{(character) Path to where the data files are stored.}

\item{data.table}{(character) Names of tabular data. Names of non-tabular data are
entered as arguments to `make_eml`.}

\item{x}{(named list) Alternative input/output to `EMLassemblyline` functions. 
Use \code{template_arguments} to create `x`.}

\item{write.file}{(logical) Write template files to `path`.}
}
\value{
\itemize{
        \item{`attributes_*.txt`} Template(s) for data table attribute 
        metadata, populated with some automatically extracted metadata 
        content from the data files.
        \item{`custom_units.txt`} Template for defining non-standard units 
        used in the `attributes_*.txt` file.
        \item{If using `x`, then content of each above listed template file 
        is added to `x` under `/x/templates/`}
    }
}
\description{
Import template files for storage of core metadata. Some templates are 
    populated with content detected by automated metadata extraction 
    methods. The remainder will have to be manually entered by a human. 
    Instructions for filling out the templates are at
    \url{https://clnsmth.github.io/EMLassemblyline/articles/instructions.html}.
}
\details{
Existing templates will not be overwritten by `template_table_attributes`.
}
