% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/template_table_attributes.R
\name{template_table_attributes}
\alias{template_table_attributes}
\title{Create table attributes template}
\usage{
template_table_attributes(
      path,
      data.path = path,
      data.table = NULL,
      write.file = TRUE,
      x = NULL
    )
}
\arguments{
\item{path}{(character) Path to the directory where the attributes template will be
written.}

\item{data.path}{(character) Path to the directory containing \code{data.table}.}

\item{data.table}{(character) Name of data table. If more than one, then supply as a 
vector of character strings (e.g. 
\code{data.table = c('concentrations.csv', 'characteristics.csv')}).}

\item{x}{(named list; optional) Alternative input to 
\code{EMLassemblyline} functions. Use \code{template_arguments()} to 
create \code{x}.}

\item{write.file}{(logical; optional) Whether to write the attribute template(s) to 
\code{path}.}
}
\value{
\itemize{
        \item{\strong{attributes_*.txt} The table attribute template. A tab
        delimited table.}
        \item{\strong{custom_units.txt} Template for defining non-standard 
        units if used in \strong{attributes_*.txt}. A tab delimited table.}
        \item{If using \code{x}, then the attribute template is added to 
        \strong{/x/templates}.}
    }
}
\description{
Create the data table attributes template. Some fields are populated 
    with content detected by automated metadata extraction methods. The 
    remainder is manually entered by the user. 
    \href{https://ediorg.github.io/EMLassemblyline/articles/edit_metadata_templates.html}{Instructions for editing the template.}
}
\details{
An existing attributes template will not be overwritten by subsequent 
    calls to \code{template_table_attributes()}.
}
\examples{
# Set working directory
setwd(tempdir())

# Create data package directory "edi_250"
file.copy(
 from = system.file('/examples/edi_250', package = 'EMLassemblyline'),
 to = '.',
 recursive = TRUE
)

# View directory contents (NOTE: attributes*_.txt don't exist)
dir('./edi_250/metadata_templates')

# Template table attributes
template_table_attributes(
  path = './edi_250/metadata_templates',
  data.path = './edi_250/data_objects',
  data.table = c('decomp.csv', 'nitrogen.csv')
)

# View directory contents (NOTE: attributes*_.txt exist)
dir('./edi_250/metadata_templates')

# Rerunning template_table_attributes() does not overwrite files
template_table_attributes(
  path = './edi_250/metadata_templates',
  data.path = './edi_250/data_objects',
  data.table = c('decomp.csv', 'nitrogen.csv')
)
    
}
