% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/template_taxonomic_coverage.R
\name{template_taxonomic_coverage}
\alias{template_taxonomic_coverage}
\title{Describe biological organisms (taxa)}
\usage{
template_taxonomic_coverage(
  path,
  data.path = path,
  taxa.table,
  taxa.col,
  taxa.name.type,
  taxa.authority,
  write.file = TRUE,
  x = NULL
)
}
\arguments{
\item{path}{(character) Path to the metadata template directory.}

\item{data.path}{(character) Path to the data directory.}

\item{taxa.table}{(character) Table name containing \code{taxa.col}. If inputting more 
than one table, then supply as a list (e.g. 
\code{c('table.1', 'table.2')}).}

\item{taxa.col}{(character) Column name containing taxa names. Species binomials are
accepted. If inputting more than one table, then supply as a list
\code{c('taxa.col.table.1', 'taxa.col.table.2')}.}

\item{taxa.name.type}{(character) Taxa name type. Can be: 
\code{scientific}, \code{common}, or \code{both}.}

\item{taxa.authority}{(integer) Taxonomic authorities to resolve to as an ordered numeric vector 
of ID's decreasing in preference. See the list of supported authorities 
with \code{taxonomyCleanr::view_taxa_authorities()}. Columns 
"resolve_sci_taxa", and "resolve_comm_taxa" list authorites supporting 
scientific and common searches, respectively.}

\item{write.file}{(logical; optional) Whether to write the template file.}

\item{x}{(named list; optional) Alternative input to 
\code{template_taxonomic_coverage()}. Use \code{template_arguments()} 
to create \code{x}.}
}
\value{
\item{taxonomic_coverage}{Columns:
    \itemize{
    \item{taxa_raw: Taxon name as it occurs in the data and as it will be listed in the metadata if no value is listed under the name_resolved column. Can be single word or species binomial.}
    \item{name_type: Type of name. Can be "scientific" or "common".}
    \item{name_resolved: Taxons name as found in an authority system.}
    \item{authority_system: Authority system in which the taxa’s name was found. Can be: "\href{https://www.itis.gov/}{ITIS}", "\href{http://www.marinespecies.org/}{WORMS}", "or "\href{https://www.gbif.org/dataset/d7dddbf4-2cf0-4f39-9b2a-bb099caae36c}{GBIF}".}
    \item{authority_id: Taxa’s identifier in the authority system (e.g. 168469).}
    }
}
}
\description{
Describes biological organisms (taxa) occuring in the data and helps resolve them to authority systems. If matches can be made, then the full taxonomic hierarchy of scientific and common names are automatically rendered in the final EML metadata. This enables future users to search on any taxonomic level of interest across data packages in data repositories.
}
\details{
\code{template_taxonomic_coverage()} searches the most preferred 
    taxonomic authority for all unique taxa listed in \code{taxa.col} 
    returning the authority name and corresponding taxa identifier for 
    direct matches (no fuzzy searching), then the next most preferred 
    taxonomic authority is search for taxa that have not yet been resolved. 
    This process repeats for subsequently listed authorities. "NA" is 
    returned when an authority match is not made.
    
    When taxonomic_coverage.txt is passed to \code{make_eml()}, the 
    authority information is used to get the hierarchical rank names of 
    resolved taxa and rendered into the "taxonomicCoverage" element of EML.
    
    Character encoding of metadata extracted directly from the tables are 
    converted to UTF-8 via \code{enc2utf8()}.
}
\examples{
\dontrun{
# Set working directory
setwd("/Users/me/Documents/data_packages/pkg_260")

# For a table containing taxa, try resolving scientific names to ITIS, and if no matches can be made, then try WORMS
# Template taxonomic coverage
template_taxonomic_coverage(
  path = "./metadata_templates",
  data.path = "./data_objects",
  taxa.table = "decomp.csv",
  taxa.col = "taxa",
  taxa.authority = c(3,9),
  taxa.name.type = "scientific")
}

}
