% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_templates.R
\name{validate_abstract}
\alias{validate_abstract}
\title{Validate the abstract template}
\usage{
validate_abstract(x)
}
\arguments{
\item{x}{(named list) The data and metadata object returned by 
\code{template_arguments()}.}
}
\value{
(character or NULL) A character vector of issues if any were found, 
    otherwise NULL.
}
\description{
Validate the abstract template
}
\details{
Checks performed by this function:
    \itemize{
        \item{Abstract is not empty}
    }
}
\keyword{internal}
