% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_arguments.R
\name{validate_arguments}
\alias{validate_arguments}
\title{Validate arguments of `EMLassemblyline` functions}
\usage{
validate_arguments(x)
}
\arguments{
\item{x}{(list) A named list of arguments and corresponding values created by a
call to `as.list(environment())` within the function whos arguments are
being validated.}
}
\description{
Validate input arguments to `EMLassemblyline` functions.
}
