% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_keywords.R
\name{validate_keywords}
\alias{validate_keywords}
\title{validate_keywords}
\usage{
validate_keywords(path, cv)
}
\arguments{
\item{path}{(character) Path to the directory containing keywords.txt.}

\item{cv}{(character) A controlled vocabulary to search. Valid options are:
\itemize{
    \item{lter} - The LTER Controlled Vocabulary (http://vocab.lternet.edu/vocab/vocab/index.php)
}}
}
\value{
An updated version of keywords.txt with the controlled vocabulary name 
    listed along side keywords for which a direct match was found. For each
    keyword that could not be matched, a set of options are listed in the 
    RStudio Console Window that the user must manually select from. The 
    results of this selection process are written to keywords.txt.
}
\description{
Check the validity of terms listed in keywords.txt against controlled 
    vocabularies, and receive suggestions for improving non-aligned terms.
}
