% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_personnel.R
\name{validate_personnel}
\alias{validate_personnel}
\title{Validate personnel}
\usage{
validate_personnel(x = personinfo)
}
\arguments{
\item{x}{The data frame created from reading personnel.txt.}
}
\value{
If incongruence is found an error is returned, else a valid data frame
    for inputs to the personnel EML making functions.
}
\description{
This function checks the personnel table for required information and 
    adjusts content to meet input requirements of the personnel EML making 
    functions.
}
