% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_templates.R
\name{validate_provenance}
\alias{validate_provenance}
\title{Validate the provenance template}
\usage{
validate_provenance(x)
}
\arguments{
\item{x}{(list) The data and metadata object returned by 
\code{template_arguments()}.}
}
\value{
\item{issues}{(character) Descriptions of issues found in the template}
    \item{x}{(list) A potentially modified \code{x} if relevant issues were
    found}
}
\description{
Validate the provenance template
}
\details{
Checks performed by this function:
    \itemize{
        \item{Template column names are correct}
        \item{systemID is supported}
        \item{dataPackageID and systemID pair resolves to resource metadata}
        \item{All fields are complete, when dataPackageID and systemID pair is not present}
        \item{A creator and contact is listed for each resource, when dataPackageID and systemID pair is not present}
        \item{Title is present, when dataPackageID and systemID pair is not present}
        \item{URL present, when dataPackageID and systemID pair is not present}
    }
    
    Checks are grouped by required and optional criteria. If any required
    checks fail, then the entire template is removed from \code{x}.
}
