% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_templates.R
\name{validate_templates}
\alias{validate_templates}
\title{Validate metadata template content}
\usage{
validate_templates(
      fun.name,
      fun.args
    )
}
\arguments{
\item{fun.name}{(character) Function name passed to `validate_x` with 
`as.character(match.call()[[1]])`.}

\item{fun.args}{(named list) Function arguments and values passed to `validate_x` with 
`as.list(environment())`.}
}
\description{
Validate the content of `EMLassembline` metadata templates.
}
\details{
Validation checks are function specific.
}
