% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_attributes.R
\name{write_attributes}
\alias{write_attributes}
\title{Write attributes from input metadata}
\usage{
write_attributes(path)

    Run this function whenever any of your 
    \emph{datasetname_datatablename_attributes_draft.xlsx} or referenced 
    data tables have changed.
}
\arguments{
\item{path}{A path to the dataset working directory containing the 
completed \emph{datasetname_datatablename_attributes_draft.xlsx}, and
\emph{eml_configuration.R} files as well as the referenced data table.}
}
\value{
A file in the dataset working directory titled 
    \emph{datasetname_datatablename_attributes.xlsx} containing attribute 
    information called upon by the functions \emph{write_factors} and 
    \emph{create_eml}.
}
\description{
A function to compile attributes for \emph{write_factors} 
    and \emph{create_eml}.
}
\details{
This function compiles information from 
    \emph{datasetname_datatablename_attributes_draft.xlsx}, extracts 
    attribute number types and minimum and maximum values from numeric 
    attribues of your data table, and formats this meta data for the 
    functions \emph{write_factors} and \emph{create_eml}.
}
\seealso{
\code{\link{copy_templates}} to copy metadata templates to the 
    dataset working directory.

\code{\link{run_guide}} for guidance on completing the template 
    files.

\code{\link{write_factors}} to create the factors table if the 
    attributes table contains factors.

\code{\link{extract_geocoverage}} to extract detailed geographic 
    coordinates of sampling sites.

\code{\link{create_eml}} to write the EML file.
}
