% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_attributes.R
\name{write_attributes}
\alias{write_attributes}
\title{Write data table attributes}
\usage{
write_attributes(path)
}
\arguments{
\item{path}{A path to the working directory containing files required by the EMLtools
package. Run the function run_guide() for instructions on how to populate this working
directory with the requisite files.}
}
\value{
A .xlsx file in the specified working directory containing an approximated set
of attributes for a specified data table. The user is prompted to refine this table.
}
\description{
A function for writting data table attributes.
}
\seealso{
\code{\link{run_guide}} for guidance on setting up the working directory

\code{\link{eml_configuration}} for configuring eml

\code{\link{write_factors}} for writing data factors

\code{\link{create_eml}} for creating EML
}
