% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_factors.R
\name{write_factors}
\alias{write_factors}
\title{Write data table factors}
\usage{
write_factors(path)
}
\arguments{
\item{path}{A path to the working directory containing files required by the EMLtools
package. Run the function run_guide() for instructions on how to populate this working
directory with the requisite files.}
}
\value{
A .xlsx file in the specified working directory containing an approximated set
of factors based on the user supplied attributes.xlsx. The user is prompted to refine
this table.
}
\description{
A function for writting data table factors.
}
\seealso{
\code{\link{run_guide}} for guidance on setting up the working directory

\code{\link{write_attributes}} for writing data attributes

\code{\link{create_eml}} for creating EML

\code{\link{copy_templates}} for creating template files
}
